### 2025-04-23 #8093 @samhforbes
Hi @vinayakdasgupta due to the relatively small size of this package I'm going to ask our editors to assess whether it is in scope for JOSS, particularly our requirements for substantial scholarly effort. In the meantime you can see you have a pdf issue and missing license that it would be good to fix.
