### 2025-04-23 f04ba9d
fixed paper.md, added paper.bib, and LICENSE
