+++ paper.md
@@ -14,20 +14,18 @@ authors:
 affiliations:
   - name: Shiv Nadar Institute of Eminence
     index: 1
-date: 2025-04-23
-bibliography: paper.bib
+date: "2025-04-23"
 ---
 
 # Summary
 
 *anvay* is a web-based tool for topic modelling in bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA)—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, *anvay* presents model results as provisional surfaces to be read, questioned, and restructured by the user.
 
-Designed for literary, journalistic, and historical corpora in bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation [@chakrabarty-etal-2017-context; @alam2021review], frequency filtering, and stopword pruning. The outputs—ranging from topic-word networks to document-level previews—are rendered with clarity and designed to scaffold close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
+Designed for literary, journalistic, and historical corpora in bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation, frequency filtering, and stopword pruning. The outputs—ranging from topic-word networks to document-level previews—are rendered with clarity and designed to scaffold close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
 
 Beyond its technical function, *anvay* is an intervention in how we teach and understand computational methods in low-resource contexts. It centres the reader, not the algorithm. Its value lies not in automation, but in the space it creates for interpretation, curation, and the possibility of collaborative meaning-making.
 
 The source code and full documentation are available at: https://github.com/vinayakdasgupta/anvay
-
 
 # Statement of Need
 
@@ -41,7 +39,7 @@ Built in Python using Flask, Gensim, and standard data visualisation libraries, 
 
 - **Corpus upload and preprocessing**: Users may upload up to 800 `.txt` files in a single run. Preprocessing options include standard and custom stopword filtering, optional stemming or lemmatisation, and frequency-based token pruning.
 - **Topic model training**: Users can adjust passes, iterations, alpha, eta, chunk size, and other hyperparameters. Models are trained on the server and stored temporarily for display.
-- **Language-specific processing**: Bengali tokenisation is handled with custom rules to prevent malformed segmentation. Lemma support is drawn from publicly available datasets [@chakrabarty-etal-2017-context].
+- **Language-specific processing**: Bengali tokenisation is handled with custom rules to prevent malformed segmentation. Lemma support is drawn from publicly available datasets.
 - **Visual outputs**: Model results are rendered via multiple interactive modules:
   - PCA-reduced topic plots (Plotly)
   - Topic-word heatmaps (Seaborn)
@@ -59,13 +57,11 @@ In pedagogical contexts, *anvay* functions as a conceptual primer. It prompts st
 
 To support this, *anvay* includes extensive web-based documentation—not just technical, but interpretive. Each section guides users through corpus preparation, parameter tuning, and result analysis, with annotated examples and embedded visual references. The documentation foregrounds conceptual understanding: users are encouraged to read models critically, experiment with settings, and reflect on how computational structure intersects with thematic interpretation. It is embedded directly in the interface and designed for both classroom and independent study.
 
-
 # Comparison with Existing Tools
 
[…]
