### 2025-04-23 #8093 @vinayakdasgupta
Thanks for the update. I've now addressed the two technical issues:

- PDF generation issue resolved: `paper.md` was revised and a `paper.bib` file added.
- License added: A valid MIT `LICENSE` file is now present in the repository root.

Please let me know if anything else is needed on my end. Happy to provide additional context on the scholarly contribution if helpful. Looking forward to the next steps.
