+++ paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-*anvay* is a web-based tool for topic modelling in bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA)—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, *anvay* presents model results as provisional surfaces to be read, questioned, and restructured by the user.
+*anvay* is a web-based tool for topic modelling in bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA) [@blei2003latent]—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, *anvay* presents model results as provisional surfaces to be read, questioned, and restructured by the user.
 
 Designed for literary, journalistic, and historical corpora in bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation, frequency filtering, and stopword pruning. The outputs—ranging from topic-word networks to document-level previews—are rendered with clarity and designed to scaffold close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
 
@@ -30,7 +30,7 @@ The source code and full documentation are available at: https://github.com/vina
 
 # Statement of Need
 
-There are few interpretive modelling tools for bengali, and fewer still that support pedagogical engagement. Most existing frameworks for LDA assume technical fluency, rely on scripts or notebooks, and provide minimal attention to interpretability. While backend libraries like Gensim and Mallet offer robust implementations, they do not assist users in making sense of what a topic is, how it shifts across documents, or how it might relate to thematic or conceptual questions.
+There are few interpretive modelling tools for bengali, and fewer still that support pedagogical engagement. Most existing frameworks for LDA assume technical fluency, rely on scripts or notebooks, and provide minimal attention to interpretability. While backend libraries like Gensim [@rehurek2010software] and Mallet offer robust implementations, they do not assist users in making sense of what a topic is, how it shifts across documents, or how it might relate to thematic or conceptual questions.
 
 *anvay* addresses this gap by making topic modelling visible and usable. It offers an in-browser interface with no coding requirements, and is particularly suited to students, independent researchers, and archival practitioners working in South Asian languages. It allows users to upload their own text corpora, adjust model parameters, and inspect topic boundaries through interactive, multi-modal visualisations. The tool has already been used in classroom and workshop settings to introduce interpretive topic modelling to humanities students with no programming background.
 
@@ -40,7 +40,7 @@ Built in Python using Flask, Gensim, and standard data visualisation libraries, 
 
 - **Corpus upload and preprocessing**: Users may upload up to 800 `.txt` files in a single run. Preprocessing options include standard and custom stopword filtering, optional stemming or lemmatisation, and frequency-based token pruning.
 - **Topic model training**: Users can adjust passes, iterations, alpha, eta, chunk size, and other hyperparameters. Models are trained on the server and stored temporarily for display.
-- **Language-specific processing**: Bengali tokenisation is handled with custom rules to prevent malformed segmentation. Lemma support is drawn from publicly available datasets.
[…]
