+++ paper.md
@@ -60,7 +60,7 @@ To support this, *anvay* includes extensive web-based documentation—not just t
 
 # Comparison with Existing Tools
 
-Most open-source topic modelling tools prioritise scale or coherence evaluation. *pyLDAvis*, for instance, provides excellent topic distance visualisations and relevance-based ranking, but relies on pre-tokenised, preprocessed input and assumes English-language conventions. It offers no support for Indic scripts, and its interface—though informative—is not designed for novice users or pedagogical contexts. Similarly, Voyant Tools [@voyant] provides a visually rich environment for text exploration, but lacks custom LDA configurability and does not support Bengali tokenisation or lemmatisation.
+Most open-source topic modelling tools prioritise scale or coherence evaluation. *pyLDAvis* [@pyldavis2014], for instance, provides excellent topic distance visualisations and relevance-based ranking, but relies on pre-tokenised, preprocessed input and assumes English-language conventions. It offers no support for Indic scripts, and its interface—though informative—is not designed for novice users or pedagogical contexts. Similarly, Voyant Tools [@voyant] provides a visually rich environment for text exploration, but lacks custom LDA configurability and does not support Bengali tokenisation or lemmatisation.
 
 *anvay* differs in both scope and ethos. It is built for **interpretation over evaluation**, and for **accessibility over scale**. Unlike pyLDAvis, which emphasises statistical overlap, *anvay* foregrounds **contextual previews**, **representative sentences**, and **visual topic-document mappings**—elements designed to support **reading**, not just inspection. Unlike Voyant, which is multilingual but limited in model customisation, *anvay* allows fine-tuning of LDA parameters, stopword filters, and token thresholds—all with support for Bengali scripts and morphologies. These choices reflect the tool’s commitment to low-resource contexts, and its role as both a scholarly and pedagogical infrastructure.
 
