### 2025-04-23 #8093 @vinayakdasgupta
Thanks for your note, @samhforbes, and I appreciate your consideration. While anvay may appear modest in size, I’d like to clarify how it aligns with JOSS’s scope through its software architecture, reproducibility, and pedagogical intent.

anvay is a complete research software package for topic modelling a non-Latin-script corpora, with:
- A modular LDA pipeline (parameterised and extensible, built on gensim)
- Integrated visualisation generation (plotly, bokeh, seaborn) using reusable wrappers
- Bengali-specific text preprocessing (custom stemmer, stopword handling, encoding clean-up)
- A structured interface for document-topic interpretability (representative sentences, topic-document mapping)
- Open-source licensing, example corpora, and install/run instructions

Beyond these technical aspects, anvay is accompanied by extensive and structured documentation aimed at new users, particularly students and researchers in South Asian digital humanities. The tool is designed not only for reproducible topic modelling, but also for interpretive exploration and pedagogical use. This includes:

- A full walkthrough of a real-world corpus analysis
- Clear explanations of each parameter and its effect
- A visual guide to outputs, supported by downloadable examples
- Accessibility features like dark mode, language toggles, and mobile responsiveness

While tailored to Bengali, the software is structured to support broader adaptation for other Indic languages or low-resource contexts, a contribution to the open-source research software ecosystem.

I have addressed the license and PDF issues. Please let me know if there's anything I can add to further demonstrate the scope and research significance of the tool. I look forward to hearing from the Editorial team.
