### 2025-05-20 3f2a163
added comprehensive tests at /tests
