### 2025-05-30 #8093 @vinayakdasgupta
Hi @samhforbes, thank you for the update. I  understand and appreciate the care you and the other editors are taking with the decision. I'm happy to wait and look forward to hearing from you whenever a consensus is reached.
