### 2025-07-08 e66b478
updated PAPER.md, PAPER.bib, about.html, README.md
