+++ paper.bib
@@ -7,27 +7,44 @@
   year={2003}
 }
 
-@inproceedings{rehurek2010software,
-  title={Software Framework for Topic Modelling with Large Corpora},
-  author={Řehůřek, Radim and Sojka, Petr},
-  booktitle={Proceedings of the LREC 2010 Workshop on New Challenges for NLP Frameworks},
-  year={2010}
+@inproceedings{rehurek_lrec,
+      title = {{Software Framework for Topic Modelling with Large Corpora}},
+      author = {Radim {\v R}eh{\r u}{\v r}ek and Petr Sojka},
+      booktitle = {{Proceedings of the LREC 2010 Workshop on New
+           Challenges for NLP Frameworks}},
+      pages = {45--50},
+      year = 2010,
+      month = May,
+      day = 22,
+      publisher = {ELRA},
+      address = {Valletta, Malta},
+      note={\url{http://is.muni.cz/publication/884893/en}},
+      language={English}
 }
 
 @inproceedings{hagberg2008exploring,
   title={Exploring Network Structure, Dynamics, and Function using NetworkX},
   author={Hagberg, Aric and Swart, Pieter and Schult, Daniel A.},
   booktitle={Proceedings of the 7th Python in Science Conference (SciPy)},
-  year={2008}
+  year={2008},
+  doi={10.25080/TCWV9851}
 }
 
 @inproceedings{chakrabarty-etal-2017-context,
-  title={Context Sensitive Lemmatization Using Two Successive Bidirectional Gated Recurrent Networks},
-  author={Chakrabarty, Arindam and Pandit, Om Pankaj and Garain, Utpal},
-  booktitle={Proceedings of the 55th Annual Meeting of the Association for Computational Linguistics},
-  pages={1481--1491},
-  year={2017},
-  doi={10.18653/v1/P17-1136}
+    title = "Context Sensitive Lemmatization Using Two Successive Bidirectional Gated Recurrent Networks",
+    author = "Chakrabarty, Abhisek  and
+      Pandit, Onkar Arun  and
+      Garain, Utpal",
+    editor = "Barzilay, Regina  and
+      Kan, Min-Yen",
+    booktitle = "Proceedings of the 55th Annual Meeting of the Association for Computational Linguistics (Volume 1: Long Papers)",
+    month = jul,
+    year = "2017",
+    address = "Vancouver, Canada",
+    publisher = "Association for Computational Linguistics",
+    url = "https://aclanthology.org/P17-1136/",
+    doi = "10.18653/v1/P17-1136",
+    pages = "1481--1491"
 }
 
 @article{alam2021review,
@@ -52,3 +69,14 @@
   year={2016},
   howpublished={https://voyant-tools.org/}
 }
+
+@article{pedregosa2011scikit,
+  title={Scikit-learn: Machine learning in Python},
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+  journal={Journal of Machine Learning Research},
+  volume={12},
+  pages={2825--2830},
+  year={2011},
+  publisher={JMLR. org}
+}
+
