+++ paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0009-0008-5623-1226
     affiliation: 1
 affiliations:
-  - name: Shiv Nadar Institute of Eminence
+  - name: Shiv Nadar Institution of Eminence
     index: 1
 date: "2025-04-23"
 bibliography: paper.bib
@@ -20,85 +20,57 @@ bibliography: paper.bib
 
 # Summary
 
-*anvay* is a web-based tool for topic modelling in bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA) [@blei2003latent]—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, *anvay* presents model results as provisional surfaces to be read, questioned, and restructured by the user.
+*anvay* is a web-based tool for topic modelling in bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA) [@blei2003latent]—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, anvay presents the model output to be read, interpreted, and adjusted by the user.
 
-Designed for literary, journalistic, and historical corpora in bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation, frequency filtering, and stopword pruning. The outputs—ranging from topic-word networks to document-level previews—are rendered with clarity and designed to scaffold close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
+Designed for literary, journalistic, and historical corpora in bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation, frequency filtering, and stopword pruning. The outputs, ranging from topic-word networks to document-level previews, are rendered with clarity and designed to enable close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
 
-Beyond its technical function, *anvay* is an intervention in how we teach and understand computational methods in low-resource contexts. It centres the reader, not the algorithm. Its value lies not in automation, but in the space it creates for interpretation, curation, and the possibility of collaborative meaning-making.
-
-The source code and full documentation are available at: https://github.com/vinayakdasgupta/anvay
+Beyond its technical function, *anvay* is an intervention in how we teach and understand computational methods within the humanities in low-resource contexts. 
 
 # Statement of Need
 
-There are few interpretive modelling tools for bengali, and fewer still that support pedagogical engagement. Most existing frameworks for LDA assume technical fluency, rely on scripts or notebooks, and provide minimal attention to interpretability. While backend libraries like Gensim [@rehurek2010software] and Mallet offer robust implementations, they do not assist users in making sense of what a topic is, how it shifts across documents, or how it might relate to thematic or conceptual questions.
-
-*anvay* addresses this gap by making topic modelling visible and usable. It offers an in-browser interface with no coding requirements, and is particularly suited to students, independent researchers, and archival practitioners working in South Asian languages. It allows users to upload their own text corpora, adjust model parameters, and inspect topic boundaries through interactive, multi-modal visualisations. The tool has already been used in classroom and workshop settings to introduce interpretive topic modelling to humanities students with no programming background.
-
-# Functionality
-
-Built in Python using Flask, Gensim, and standard data visualisation libraries, *anvay* provides the following core functionalities:
-
[…]
