### 2025-07-08 #8093 @vinayakdasgupta
Dear @abhishektiwari I have addressed the issues that you had raised:

1. I have reduced the word count of PAPER.md to 1000 words
2. I have added a CONTRIBUTING.md
3. fixed missing reference
4. added DOI

Do let me know if this is okay.
