### 2025-07-15 #8093 @vinayakdasgupta
Hi@abhishektiwari, Dr. Shanmugapriya T wrote to me earlier and expressed her willingness to review the project. However, she also suggested that we include her Ph.D. student who works on topic models and reads Bengali for this review. Would this be possible? If you could let her know, I would be very grateful. Thank you.
