### 2025-07-24 #8093 @abhishektiwari
👋 @dondealban @sagorbrur, would any of you be willing to review this submission for JOSS?  The submission under consideration is "anvay: a flask-based tool for interpretive topic modelling in Bengali".

Based on your experience in topic modelling with focus on South Asian languages, we think you might be able to provide a great review of this submission.

Thanks
Abhi
