### 2025-07-25 #8641 @editorialbot
[REVIEW]: anvay: a flask-based tool for interpretive topic modelling in Bengali
**Submitting author:** @vinayakdasgupta (<a href="http://orcid.org/0009-0008-5623-1226">Vinayak Das Gupta</a>)
**Repository:** https://github.com/vinayakdasgupta/anvay
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
