+++ uploads/ছত্রিশ.txt
@@ -1,0 +1,55 @@
+শীতের রোদ্দুর।
+সোনা - মেশা সবুজের ঢেউ
+স্তম্ভিত হয়ে আছে সেগুন বনে।
+বেগ্ নি - ছায়ার ছোঁওয়া - লাগা
+ঝুরি - নামা বৃদ্ধ বট
+ডাল মেলেছে রাস্তার ও পার পর্যন্ত।
+ফলসাগাছের ঝরা পাতা
+হঠাৎ হাওয়ায় চমকে
+বেড়ায় উড়ে
+ধুলোর সাঙাত হয়ে।
+কাজ - ভোলা এই দিন
+উধাও বলাকার মতো
+লীন হয়ে চলেছে নিঃসীম নীলিমায়।
+ঝাউগাছের
+মর্মরধ্বনিতে মিশে
+মনের
+মধ্যে এই কথাটি উঠছে বেজে,
+‘ আমি
+আছি। '
+কুয়োতলার কাছে
+সামান্য ওই আমের গাছ ;
+সারা বছর ও থাকে আত্মবিস্মৃত,
+বনের সাধারণ সবুজের আবরণে
+ও থাকে ঢাকা।
+এমন সময় মাঘের শেষে
+হঠাৎ মাটির নীচে
+শিকড়ে শিকড়ে
+তার শিহর লাগে,
+শাখায় শাখায় মুকুলিত হয়ে ওঠে বাণী—
+‘ আমি আছি '।
+চন্দ্রসূর্যের আলো আপন ভাষায়
+স্বীকার করে তার সেই ভাষা।
+শেষ সপ্তক
+অলস মনের শিয়রে দাঁড়িয়ে
+হাসেন অন্তর্যামী,
+হঠাৎ দেন ঠেকিয়ে সোনার কাঠি
+প্রিয়ার মুগ্ধ চোখের দৃষ্টি দিয়ে,
+কবির
+গানের সুর দিয়ে,
+তখন যে - আমি ধূলিধূসর সামান্য
+দিনগুলির
+মধ্যে
+মিলিয়ে ছিল,
+সে দেখা দেয় এক
+নিমেষের অসমান্য আলোকে।
+সে - সব দুর্মূল্য নিমেষ
+কোনো
+রত্নভাণ্ডারে থেকে যায় কি না জানি নে ;
+এইটুকু জানি—
+তারা এসেছে আমার
+আত্মবিস্মৃতির মধ্যে,
+জাগিয়েছে আমার মর্মে
+বিশ্বমর্মের
+নিত্যকালের সেই বাণী
+‘ আমি আছি '।
