+++ uploads/কৃষ্ণকলি.txt
@@ -1,0 +1,41 @@
+কৃষ্ণকলি আমি তারেই বলি,
+কালো তারে বলে গাঁয়ের লোক।
+মেঘলা দিনে দেখেছিলেম মাঠে
+কালো মেয়ের কালো হরিণ - চোখ।
+ঘোমটা মাথায় ছিল না তার মোটে,
+মুক্তবেণী পিঠের ‘পরে লোটে।
+কালো? তা সে যতই কালো হোক
+দেখেছি তার কালো হরিণ - চোখ।
+ঘন মেঘে আঁধার হল দেখে
+ডাকতেছিল শ্যামল দুটি গাই,
+শ্যামা মেয়ে ব্যস্ত ব্যাকুল পদে
+কুটির হতে ত্রস্ত এল তাই।
+আকাশ - পানে হানি যুগল ভুরু
+শুনলে বারেক মেঘের গুরুগুরু।
+কালো? তা সে যতই কালো হোক
+দেখেছি তার কালো হরিণ - চোখ।
+পুবে বাতাস এল হঠাৎ ধেয়ে,
+ধানের খেতে খেলিয়ে গেল ঢেউ।
+আলের ধারে দাঁড়িয়েছিলেম একা,
+মাঠের মাঝে আর ছিল না কেউ।
+আমার পানে দেখলে কিনা চেয়ে
+আমিই জানি আর জানে সে মেয়ে।
+কালো? তা সে যতই কালো হোক
+দেখেছি তার কালো হরিণ - চোখ।
+এমনি করে কালো কাজল মেঘ
+জ্যৈষ্ঠ মাসে আসে ঈশান কোণে।
+এমনি করে কালো কোমল ছায়া
+ক্ষণিকা
+আষাঢ় মাসে নামে তমাল - বনে।
+এমনি করে শ্রাবণ - রজনীতে
+হঠাৎ খুশি ঘনিয়ে আসে চিতে।
+কালো? তা সে যতই কালো হোক
+দেখেছি তার কালো হরিণ - চোখ।
+কৃষ্ণকলি আমি তারেই বলি,
+আর যা বলে বলুক অন্য লোক।
+দেখেছিলেম ময়নাপাড়ার মাঠে
+কালো মেয়ের কালো হরিণ - চোখ।
+মাথার ' পরে দেয় নি তুলে বাস,
+লজ্জা পাবার পায় নি অবকাশ।
+কালো? তা সে যতই কালো হোক
+দেখেছি তার কালো হরিণ - চোখ।
