+++ uploads/উনিশ.txt
@@ -1,0 +1,54 @@
+তখন বয়স ছিল কাঁচা ;
+কতদিন মনে মনে এঁকেছি নিজের
+ছবি,
+বুনো ঘোড়ার পিঠে সওয়ার,
+জিন নেই, লাগাম নেই,
+ছুটেছি ডাকাত-হানা মাঠের মাঝখান দিয়ে
+ভরসন্ধ্যেবেলায় ;
+ঘোড়ার খুরে উড়ছে ধুলো
+ধরণী যেন পিছু ডাকছে আঁচল দুলিয়ে।
+আকাশে সন্ধ্যার প্রথম তারা
+দূরে মাঠের সীমানায়
+দেখা যায়
+একটিমাত্র ব্যগ্র বিরহী আলো একটি কোন্ ঘরে
+নিদ্রাহীন প্রতীক্ষায়।
+যে ছিল ভাবীকালে
+আগে হতে মনের মধ্যে
+ফিরছিল তারি আবছায়া,
+যেমন ভাবী আলোর আভাস আসে
+ভোরের প্রথম কোকিল-ডাকা অন্ধকারে।
+তখন অনেকখানি সংসার ছিল অজানা,
+আধো
+জানা।
+তাই অপরূপের রাঙা
+রঙটা
+মনের দিগন্ত
+রেখেছিলে রাঙিয়ে ;
+আসন্ন ভালোবাসা
+এনেছিল অঘটন-ঘটনার স্বপ্ন।
+তখন ভালোবাসার যে
+কল্পরূপ ছিল মনে
+তার সঙ্গে
+মহাকাব্যযুগের
+দুঃসাহসের আনন্দ ছিল মিলিত।
+এখন অনেক খবর পেয়েছি জগতের,
+মনে ঠাওরেছি
+শেষ সপ্তক
+সংসারের অনেকটাই মার্কামারা
+খবরের
+মালখানা।
+মনের
+রসনা থেকে
+অজানার স্বাদ গেছে মরে,
+অনুভবে পাই নে
+ভালোবাসায় সম্ভবের
+মধ্যে
+নিয়তই অসম্ভব,
+জানার মধ্যে অজানা,
+কথার মধ্যে রূপকথা
+।
+ভুলেছি প্রিয়ার মধ্যে আছে সেই নারী,
+যে থাকে সাত সমুদ্রের পারে,
+সেই নারী আছে
+বুঝি মায়ার ঘুমে,
+যার জন্যে খুঁজতে হবে সোনার কাঠি।
