+++ uploads/বিস্মরণ.txt
@@ -1,0 +1,41 @@
+মনে আছে কার দেওয়া সেই ফুল?
+সে ফুল যদি শুকিয়ে গিয়ে থাকে
+তবে তারে সাজিয়ে রাখাই ভুল—
+মিথ্যে কেন কাঁদিয়ে রাখ তাকে।
+ধুলায় তারি শান্তি তারি গতি,
+এই সমাদর কোরো তাহার প্রতি—
+সময় যখন গেছে তখন তারে
+ভুলো একেবারে।
+মাঘের শেষে নাগকেশরের ফুলে
+আকাশে বয় মন-হারানো হাওয়া;
+বনের বক্ষ উঠেছে আজ দুলে,
+চামেলি ওই কার যেন পথ-চাওয়া।
+ছায়ায় ছায়ায় কাদের কানাকানি,
+চোখে-চোখে নীরব জানাজানি—
+এ উৎসবে শুকনো ফুলের লাজ
+ঘুচিয়ে দিয়ো আজ।
+যদি-বা তার ফুরিয়ে থাকে বেলা,
+মনে জেনো দুঃখ তাহে নাই;
+করেছিল ক্ষণকালের খেলা,
+পেয়েছিল ক্ষণকালের ঠাঁই।
+অলকে সে কানের কাছে দুলি
+বলেছিল নীরব কথাগুলি,
+গন্ধ তাহার ফিরেছে পথ ভুলে
+তোমার এলোচুলে।
+সেই মাধুরী আজ কি হবে ফাঁকি।
+লুকিয়ে সে কি রয় নি কোনোখানে।
+কাহিনী তার থাকবে না আর বাকি
+পূরবী
+কোনো স্বপ্নে, কোনো গন্ধে গানে?
+আরেক দিনের বনচ্ছায়ায় লিখা
+ফিরবে না কি তাহার মরীচিকা।
+অশ্রুতে তার আভাস দিবে নাকি
+আরেক দিনের আঁখি।
+নাহয় তাও লুপ্ত যদিই হয়,
+তার লাগি শোক সেও তো সেই পথে।
+এ জগতে সদাই ঘটে ক্ষয়,
+ক্ষতি তবু হয় না কোনোমতে।
+শুকিয়ে-পড়া পুষ্পদলের ধূলি
+এ ধরণী যায় যদি বা ভুলি—
+সেই ধুলারই বিস্মরণের কোলে
+নতুন কুসুম দোলে।

[…]
