+++ sample_corpus/রাজা.txt
@@ -1,0 +1,1355 @@
+অন্ধকার ঘর
+রানী সুদর্শনা ও তাঁহার দাসী সুরঙ্গমা
+সুদর্শনা। আলো, আলো কই। এ ঘরে কি একদিনও আলো জ্বলবে না।
+সুরঙ্গমা। রানীমা, তোমার ঘরে-ঘরেই তো আলো জ্বলছে— তার থেকে সরে আসবার জন্যে কি একটা ঘরেও অন্ধকার রাখবে না।
+সুদর্শনা। কোথাও অন্ধকার কেন থাকবে।
+সুরঙ্গমা। তা হলে যে আলোও চিনবে না, অন্ধকারও চিনবে না।
+সুদর্শনা। তুই যেমন এই অন্ধকার ঘরের দাসী তেমনি তোর অন্ধকারের মতো কথা, অর্থই বোঝা যায় না। বল্ তো এ ঘরটা আছে কোথায়। কোথা দিয়ে এখানে আসি, কোথা দিয়ে বেরোই, প্রতিদিনই ধাঁদা লাগে।
+সুরঙ্গমা। এ ঘর মাটির আবরণ ভেদ করে পৃথিবীর বুকের মাঝখানে তৈরি। তোমার জন্যেই রাজা বিশেষ করে করেছেন!
+সুদর্শনা। তাঁর ঘরের অভাব কী ছিল যে এই অন্ধকার ঘরটা বিশেষ করে করেছেন!
+সুরঙ্গমা। আলোর ঘরে সকলেরই আনাগোনা— এই অন্ধকারে কেবল একলা তোমার সঙ্গে মিলন।
+সুদর্শনা। না না, আমি আলো চাই— আলোর জন্যে অস্থির হয়ে আছি। তোকে আমি আমার গলার হার দেব যদি এখানে একদিন আলো আনতে পারিস।
+সুরঙ্গমা। আমার সাধ্য কী মা— যেখানে তিনি অন্ধকার রাখেন আমি সেখানে আলো জ্বালব!
+সুদর্শনা। এত ভক্তি তোর! অথচ শুনেছি, তোর বাপকে রাজা শাস্তি দিয়েছেন। সে কি সত্যি।
+সুরঙ্গমা। সত্যি। বাবা জুয়ো খেলত। রাজ্যের যত যুবক আমাদের ঘরে জুটত—মদ খেত আর জুয়ো খেলত।
+সুদর্শনা। তুই কী করতিস।
+সুরঙ্গমা। মা, তবে সব শুনেছ। আমি নষ্ট হবার পথে গিয়েছিলমুম। বাবা ইচ্ছে করেই আমাকে সে পথে দাঁড় করিয়েছিলেন। আমার মা ছিল না।
+সুদর্শনা। রাজা যখন তোর বাপকে নির্বাসিত করে দিলেন তখন তোর রাগ হয় নি?
+সুরঙ্গমা। খুব রাগ হয়েচিল — ইচ্ছে হয়েছিল, কেউ যদি রাজাকে মেরে ফেলে তো বেশ হয়।
+সুদর্শনা। রাজা তোর বাপের কাছ থেকে ছাড়িয়ে এনে কোথায় রাখলেন?
+সুরঙ্গমা। কোথায় রাখলেন কে জানে। কিন্তু কী কষ্ট গেছে! আমাকে যেন ছুঁচ ফোটাত, আগুনে পোড়াত।
+সুদর্শনা। কেন, তোর এত কষ্ট কিসের ছিল।
[…]
