+++ sample_corpus/মেঘ ও রৌদ্র.txt
@@ -1,0 +1,189 @@
+প্রথম পরিচ্ছেদ
+পূর্বদিন বৃষ্টি হইয়া গিয়াছে। আজ ক্ষান্তবর্ষণ প্রাতঃকাল ম্লান রৌদ্র ও খণ্ড মেঘে মিলিয়া পরিপক্কপ্রায় আউশ ধানের ক্ষেত্রের উপর পর্যায়ক্রমে আপন আপন সুদীর্ঘ তুলি বুলাইয়া যাইতেছিল; সুবিস্তৃত শ্যাম চিত্রপট একবার আলোকের স্পর্শে উজ্জ্বল পাণ্ডুবর্ণ ধারণ করিতেছিল আবার পরক্ষণেই ছায়াপ্রলেপে গাঢ় স্নিগ্ধতায় অঙ্কিত হইতেছিল।
+যখন সমস্ত আকাশরঙ্গভূমিতে মেঘ এবং রৌদ্র, দুইটি মাত্র অভিনেতা, আপন আপন অংশ অভিনয় করিতেছিল তখন নিম্নে সংসাররঙ্গভূমিতে কত স্থানে কত অভিনয় চলিতেছিল তাহার আর সংখ্যা নাই।
+আমরা যেখানে একটি ক্ষুদ্র জীবননাট্যের পট উত্তোলন করিলাম সেখানে গ্রামে পথের ধারে একটি বাড়ি দেখা যাইতেছে। বাহিরের একটি মাত্র ঘর পাকা, এবং সেই ঘরের দুই পার্শ্ব দিয়া জীর্ণপ্রায় ইষ্টকের প্রাচীর গুটিকতক মাটির ঘর বেষ্টন করিয়া আছে। পথ হইতে গরাদের জানলা দিয়া দেখা যাইতেছে, একটি যুবাপুরুষ খালি গায়ে তক্তপোশে বসিয়া বামহস্তে ক্ষণে ক্ষণে তালপাতার পাখা লইয়া গ্রীষ্ম এবং মশক দূর করিবার চেষ্টা করিতেছেন এবং দক্ষিণ হস্তে বই লইয়া পাঠে নিবিষ্ট আছেন।
+বাহিরে গ্রামের পথে একটি ডুরে-কাপড়-পরা বালিকা আঁচলে গুটিকতক কালোজাম লইয়া একে একে নিঃশেষ করিতে করিতে উক্ত গরাদে-দেওয়া জানলার সম্মুখ দিয়া বারংবার যাতায়াত করিতেছিল। মুখের ভাবে স্পষ্টই বোঝা যাইতেছিল, ভিতরে যে মানুষটি তক্তপোশে বসিয়া বই পড়িতেছে তাহার সহিত বালিকার ঘনিষ্ঠ পরিচয় আছে— এবং কোনোমতে সে তাহার মনোযোগ আকর্ষণপূর্বক তাহাকে নীরবে অবজ্ঞাভরে জানাইয়া যাইতে চাহে যে, ‘সম্প্রতি কালোজাম খাইতে আমি অত্যন্ত ব্যস্ত আছি, তোমাকে আমি গ্রাহ্যমাত্র করি না।’
[…]
