+++ sample_corpus/দৃষ্টিদান.txt
@@ -1,0 +1,129 @@
+শুনিয়াছি,আজকাল অনেক বাঙালির মেয়েকে নিজের চেষ্টায় স্বামী সংগ্রহ করিতে হয়। আমিও তাই করিয়াছি,কিন্তু দেবতার সহায়তায়। আমি ছেলেবেলা হইতে অনেক ব্রত এবং অনেক শিবপূজা করিয়াছিলাম।
+আমার আটবৎসর বয়স উত্তীর্ণ না হইতেই বিবাহ হইয়া গিয়াছিল। কিন্তু পূর্বজন্মের পাপবশত আমি আমার এমন স্বামী পাইয়াও সম্পূর্ণ পাইলাম না। মা ত্রিনয়নী আমার দুইচক্ষু লইলেন। জীবনের শেষমুহূর্ত পর্যন্ত স্বামীকে দেখিয়া লইবার সুখ দিলেন না।
+বাল্যকাল হইতেই আমার অগ্নিপরীক্ষার আরম্ভ হয়। চোদ্দবৎসর পার না হইতেই আমি একটি মৃতশিশু জন্ম দিলাম;নিজেও মরিবার কাছাকাছি গিয়াছিলাম কিন্তু যাহাকে দুঃখভোগ করিতে হইবে সে মরিলে চলিবে কেন। যে দীপ জ্বলিবার জন্য হইয়াছে তাহার তেল অল্প হয় না; রাত্রিভোর জ্বলিয়া তবে তাহার নির্বাণ।
+বাঁচিলাম বটে কিন্তু শরীরের দুর্বলতায়,মনের খেদে,অথবা যে কারণেই হউক,আমার চোখের পীড়া হইল।
+আমার স্বামী তখন ডাক্তারি পড়িতেছিলেন। নূতন বিদ্যাশিক্ষার উৎসাহবশত চিকিৎসা করিবার সুযোগ পাইলে তিনি খুশি হইয়া উঠিতেন। তিনি নিজেই আমার চিকিৎসা আরম্ভ করিলেন।
+দাদা সে বছর বি.এল.দিবেন বলিয়া কালেজে পড়িতেছিলেন। তিনি একদিন আসিয়া আমার স্বামীকে কহিলেন,“করিতেছ কী। কুমুর চোখ দুটো যে নষ্ট করিতে বসিয়াছ। একজন ভালো ডাক্তার দেখাও।”
+আমার স্বামী কহিলেন,“ভালো ডাক্তার আসিয়া আর নূতন চিকিৎসা কী করিবে। ওষুধপত্র তো সব জানাই আছে।”
+দাদা কিছু রাগিয়া কহিলেন,“তবে তো তোমার সঙ্গে তোমাদের কালেজের বড়োসাহেবের কোনো প্রভেদ নাই।”
+স্বামী বলিলেন,“আইন পড়িতেছ,ডাক্তারির তুমি কি বোঝ। তুমি যখন বিবাহ করিবে তখন তোমার স্ত্রীর সম্পত্তি লইয়া যদি কখনো মকদ্দমা বাধে তুমি কি আমার পরামর্শমতো চলিবে।”
[…]

[…]
