### 2025-08-27 b3645e6
added screenshots
Added screenshots to README.md for clarity
