+++ results/topic_words.txt
@@ -1,10 +1,5 @@
-Topic 0: চরণে, আছ, স্মৃতি, বয়স, কাঠি, ধাওয়া, বিলাপ, কহিলে, বেদনা, যারে
-Topic 1: বেয়ে, স্বচ্ছ, জাত, ঘট, যা, রাঙা, লতাবিতান, জাগল, সবুজ, নত
-Topic 2: প্রদীপ, বটে, দেব, অন্ধকার, গড়া, ডাকি, রাতি, নেইকো, চাহিবে, মন্ত্র
-Topic 3: রাখি, ধরা, শেষে, লুকায়ে, ঘটা, দিক, নব, ভুলিসেই, ধরণীর, দোলে
-Topic 4: পলকে, বর, অতি, ভুলেযাওয়া, তপ্ত, পাত, বহে, বেড়ায়, নাচন, নৃত্য
-Topic 5: মূল্য, দান, কোমল, নেশা, সন্ধ্যা, শান্ত, ক্ষতি, ম্লান, তটের, হাসি
-Topic 6: লও, বোঝে, সৃষ্টি, গাহে, রহস্য, মায়া, পূজা, দীপ্ত, ধায়, খসে
-Topic 7: খাওয়া, পাড়ার, পূজা, খানে, মেঘে, পড়েযার, হাওয়ার, পেলে, ছিন্ন, পালায়
-Topic 8: রহে, আহ্বান, বিশ্ব, রক্ত, উদ্, সংশয়, মিথ্যা, আসি, সংসার, জেনো
-Topic 9: দল, যাহা, ডানার, বাণীর, ছায়ার, চরম, নিবিড়, আপনারি, বেদনার, আনা
+Topic 0: যাহা, দল, এইসব, রাখি, খসে, ডানার, নাকো, শূন্য, ঢেউয়ের, বসন্ত
+Topic 1: প্রশ্ন, অজানা, কেউ, ঘাসে, প্রদীপ, পরিচয়, চার, নবীন, তারায়, করব
+Topic 2: রহে, সংসার, ধন, লাগি, উদ্, সন্ধ্যা, জেনো, বিশ্ব, উৎসব, আহ্বান
+Topic 3: কচি, বর, বেয়ে, চুপ, আনা, নিখিল, ভরি, পড়েযার, মেঘে, কাঁপন
+Topic 4: পাড়ার, খাওয়া, পলকে, মধুর, বহে, আমের, নৃত্য, গুরু, ছুঁয়ে, ছোঁওয়া

[…]
