+++ README.md
@@ -62,11 +62,26 @@ anvay includes a fully integrated documentation panel accessible from the interf
 
 ## Screenshots
 
-![Upload interface](docs/img/anvay_upload.PNG)
-![Results interface](docs/img/anvay_results.PNG)
-![Visualization interface](docs/img/anvay_viz.PNG)
-![Documentation interface](docs/img/anvay_documentation.PNG)
-![Documentation interface](docs/img/anvay_documentation2.PNG)
+<details>
+  <summary>Upload interface screenshot</summary>
+  <img src="docs/img/anvay_upload.PNG" alt="Upload interface" width="200"/>
+</details>
+
+<details>
+  <summary>Documentation interface screenshot</summary>
+  <img src="docs/img/anvay_documentation.PNG" alt="Upload interface" width="200"/>
+</details>
+
+<details>
+  <summary>Visualization interface screenshot</summary>
+  <img src="docs/img/anvay_viz.PNG" alt="Upload interface" width="200"/>
+</details>
+
+<details>
+  <summary>Report interface screenshot</summary>
+  <img src="docs/img/anvay_results.PNG" alt="Upload interface" width="200"/>
+</details>
+
 
 ---
 
