### 2025-08-27 870740d
changing screenshot thumbnail width to 500px
