### 2025-08-27 9e63e13
adding open to <details> to keep screenshots expanded
