+++ results/topics.txt
@@ -1,15 +1,15 @@
 Topic 0:
-যাহা, দল, এইসব, রাখি, খসে, ডানার, নাকো, শূন্য, ঢেউয়ের, বসন্ত
+কালো, পথ, দেওয়া, যায়, কখন, সময়, মোর, এক, চেয়ে, পাওয়া
 
 Topic 1:
-প্রশ্ন, অজানা, কেউ, ঘাসে, প্রদীপ, পরিচয়, চার, নবীন, তারায়, করব
+তব, ওঠা, নাম, নাহি, কথা, করা, দেওয়া, ভরা, কর, রবে
 
 Topic 2:
-রহে, সংসার, ধন, লাগি, উদ্, সন্ধ্যা, জেনো, বিশ্ব, উৎসব, আহ্বান
+পাওয়া, মোর, কথা, ফুল, কোন, জানা, নাহি, পথ, এক, ওঠা
 
 Topic 3:
-কচি, বর, বেয়ে, চুপ, আনা, নিখিল, ভরি, পড়েযার, মেঘে, কাঁপন
+মোর, গান, বলা, জানা, ওগো, তব, মন, বাতাস, পথ, কথা
 
 Topic 4:
-পাড়ার, খাওয়া, পলকে, মধুর, বহে, আমের, নৃত্য, গুরু, ছুঁয়ে, ছোঁওয়া
+তুমি, মোর, আমায়, দেওয়া, আলো, হওয়া, দিন, মা, পাতা, লেখা
 
