+++ results/doc_topic_weights.txt
@@ -1,25 +1,25 @@
-৬.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.9101999998092651, Topic 4: 0.0
-১৩.txt: Topic 0: 0.0, Topic 1: 0.9526000022888184, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-অগ্রদূত.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.9193999767303467
-অটোগ্রাফ.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.9103000164031982, Topic 4: 0.0
-অধীরা.txt: Topic 0: 0.8833000063896179, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-অনবসর.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.9329000115394592, Topic 3: 0.0, Topic 4: 0.0
-অস্থানে.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.9577000141143799, Topic 4: 0.0
-আকাশ.txt: Topic 0: 0.9380000233650208, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-আকুল আহ্বান.txt: Topic 0: 0.0, Topic 1: 0.9496999979019165, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-আছি.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.9664999842643738, Topic 4: 0.0
-আন‍্মনা.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.946399986743927
-উনিশ.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.9269999861717224, Topic 3: 0.0, Topic 4: 0.0
-এক গাঁয়ে.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.9498000144958496
-কন্টিকারি.txt: Topic 0: 0.9423999786376953, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-কর্মফল.txt: Topic 0: 0.0, Topic 1: 0.8996000289916992, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-কৃষ্ণকলি.txt: Topic 0: 0.9186999797821045, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-গীতহীন.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.9575999975204468, Topic 4: 0.0
-চাবি.txt: Topic 0: 0.0, Topic 1: 0.9552000164985657, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0
-ছত্রিশ.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.8996000289916992
-জন্মকথা.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.9526000022888184, Topic 3: 0.0, Topic 4: 0.0
[…]
