### 2025-08-27 f872f2b
removed overview video due to display issues
