### 2025-11-25 vinayakdasgupta/anvay#2 @x-tabdeveloping
JOSS review - Feedback on software
Overall, I have a positive experience with the piece of software you have developed. It is very intuitive and easy to use. The updated documentation is excellent, and installing and running the app is basically trivial. I have some comments about the structure of the app, that I would like to have addressed or just explained to me if I'm misunderstanding something.

1. Your topic naming scheme is not exactly intuitive to follow. It is hard for users to know which topic is which unless they switch back and forth the report and the visualizations. Both BERTopic and topicwizard use a scheme, where the highest ranking 4-3 words make up the topic name, so users can relate that topic to something intuitive. Alternatively, users should be able to annotate topics, and name them. Correct me if I'm wrong, but a user cannot really see the top words for a given topic on the Visualizations page. They should preferably be able to do that.
2. Some of your plots show the same things. For instance, the heatmap and bar chart show you the same information, yet there is no visual indication of this. This might be a bit confusing for users.
3. Plotly has this feature, where it will hide some tick labels, where it doesn't have enough space. This can have the unintended effect that your plots become very hard to read unless one hovers on them. e.g. on this plot, half of the words are not visible:

4. The Topic Evolution plot confuses me. Users do not give you timestamps anywhere, nor is there any indication to where this "time" variable comes from. Is it within document? Is it for different documents in the order that they are uploaded to the app?
5. Hierarchical clustering is a nice touch, but it would be even better if one could see, what kinds of keywords the joint topics would contain, like in BERTopic (see Screenshot). You could also use the same color scheme for this graph.

6. This is by no means a big deal, but it would be a nice touch if your app were a bit more visually streamlined. As far as I can tell, you use Roboto as the main font, but Plotly uses Open Sans. This should be easy to fix. Plots have a  very large margin, despite already being inside a padded container. I would probably remove margins: `fig.update_layout(margin=dict(l=0, r=0, b=0, t=0))`. The `plotly_white` template would also fit the look and feel of the website better in my opinion.

I have no comments about the reports page, it is really nice, intuitive, and easy to use. Great job there.
