### 2025-11-25 #8641 @x-tabdeveloping
I have created [this issue](https://github.com/vinayakdasgupta/anvay/issues/2) on my feedback for the software. I will shortly have a look at the manuscript.
