### 2025-11-25 vinayakdasgupta/anvay#2 @x-tabdeveloping
One other thing: Your app does not display a loading screen while it is processing the data. Users that do not have access to the terminal, in which your application is running will not know what is happening, and might think that the button did nothing. You should probably add some loading screen, or a spinner.
