### 2025-12-03 #8641 @vinayakdasgupta
Dear @abhishektiwari, first off -- thanks for your patience in overseeing this. I have looked at all the comments that @x-tabdeveloping has provided. I am very grateful for these as I feel that they make the application much stronger. Over the course of the next few days, I shall work on addressing each concern.

With regards to the paper, I am very happy to expand on the state of the field section. However, this would mean that the word count would have to extend to about 2000-2500 words. Is that acceptable?

With regards to the example, I already have a Walkthrough in the documentation. I shall communicate this to the reviewer and see if this suffices or if they have something else in mind.

There are some other issues that the reviewer has pointed out with the visualization and labelling -- I shall sit with those and figure out the best possible recourse. I shall do this in communication with the reviewer.

Thank you again for this. Do let me know if you have any suggestions with regards to how I plan to address this.

Best,
Vinayak
