### 2025-12-11 c2171c1
Improve topic interpretability, unify visual styling, and update documentation (v1.1.1)
Implemented multiple usability and interpretability improvements for v1.1.1:

- Added top-word hover tooltips across all visualisations for clearer topic interpretation.
- Standardised global topic colour scheme across all charts.
- Updated documentation to explain that Heatmap and Bar Chart visualise the same topic–word weight matrix.
- Reduced number of displayed terms in plots to prevent hidden tick labels; added hover-based x-axis details where needed.
- Clarified Topic Evolution axis (document upload o
[…]
