+++ viz.py
@@ -12,10 +12,61 @@ from scipy.cluster.hierarchy import linkage, dendrogram
 from sklearn.cluster import AgglomerativeClustering
 from collections import Counter
 import re
+import json
+import uuid
+import html
 from utils import (
     custom_bengali_tokenize,
     split_sentences_bengali
 )
+
+# ------------------ Style / Theme constants (central) ------------------
+# Consistent font family (site font + Bengali shaping fallback)
+FONT_FAMILY = "Roboto, 'Noto Sans Bengali', 'Noto Sans', sans-serif"
+FONT_SIZE = 11
+
+# Brand/topic color palette:
+# Start with brand colour then fall back to Plotly qualitative palette.
+BRAND_COLOR = "#0D85D8"
+DEFAULT_PALETTE = px.colors.qualitative.Plotly  # reliable qualitative palette
+
+def build_topic_color_map(num_topics):
+    """
+    Return a dict mapping 'Topic 0'..'Topic N' -> hex colour.
+    Ensures Topic 0 uses BRAND_COLOR and cycles the default palette for others.
+    """
+    # create working palette copy
+    palette = DEFAULT_PALETTE.copy()
+    # ensure first colour is brand colour
+    if len(palette) > 0:
+        palette[0] = BRAND_COLOR
+    # extend palette if num_topics exceeds length by repeating
+    color_list = [palette[i % len(palette)] for i in range(num_topics)]
+    return {f"Topic {i}": color_list[i] for i in range(num_topics)}
+# -----------------------------------------------------------------------
+
+
+def get_topic_tooltips(lda_model, topn=5):
+    """
+    Return a dict mapping topic_id -> short string of topn words (joined by ', ').
+    """
+    tooltips = {}
+    # robustly obtain number of topics
+    num_topics = getattr(lda_model, "num_topics", None)
+    if num_topics is None:
+        # fallback for older gensim: use shape of topic matrix
+        try:
+            num_topics = lda_model.get_topics().shape[0]
+        except Exception:
+            num_topics = 0
+    for tid in range(num_topics):
+        try:
+            words = [w for w, _ in lda_model.show_topic(tid, topn=topn)]
+        except Exception:
+            # fallback to show_topic default behaviour
+            words = [w for w, _ in lda_model.show_topic(tid, topn=topn)]
+        tooltips[tid] = ", ".join(words)
+    return tooltips
 
 # Utility to export HTML divs (no full HTML)
 def plot_to_div(fig):
@@ -26,83 +77,544 @@ def create_interactive_scatter(lda_model):
     pca = PCA(n_components=2)
     coords = pca.fit_transform(topic_term_matrix)
 
+    tooltips = get_topic_tooltips(lda_model, topn=5)
+    labels = [f"Topic {i}" for i in range(len(coords))]
+
+    df = pd.DataFrame({
+        "x": coords[:, 0],
+        "y": coords[:, 1],
+        "label": labels,
+        "top_words": [tooltips[i] for i in range(len(coords))]
+    })
+
+    color_map = build_topic_color_map(len(coords))
+
+    # Tell px to colour by the label column and use our map
     fig = px.scatter(
-        x=coords[:, 0],
-        y=coords[:, 1],
-        text=[f"Topic {i}" for i in range(len(coords))],
-        hover_name=[f"Topic {i}" for i in range(len(coords))],
-        labels={'x': 'PCA 1', 'y': 'PCA 2'}
+        df,
+        x="x",
+        y="y",
+        text="label",
+        color="label",                      # <- important
+        color_discrete_map=color_map,
+        height=480
     )
-    fig.update_traces(marker=dict(size=12, color='indianred'), textposition='top center')
[…]
