+++ viz_old.py
@@ -1,0 +1,261 @@
+# viz.py (updated with Plotly and HTML embedding for Flask integration)
+
+import os
+import numpy as np
+import pandas as pd
+import networkx as nx
+import plotly.express as px
+import plotly.graph_objects as go
+import plotly.io as pio
+from sklearn.decomposition import PCA
+from scipy.cluster.hierarchy import linkage, dendrogram
+from sklearn.cluster import AgglomerativeClustering
+from collections import Counter
+import re
+from utils import (
+    custom_bengali_tokenize,
+    split_sentences_bengali
+)
+
+# Utility to export HTML divs (no full HTML)
+def plot_to_div(fig):
+    return pio.to_html(fig, full_html=False, include_plotlyjs='cdn')
+
+def create_interactive_scatter(lda_model):
+    topic_term_matrix = lda_model.get_topics()
+    pca = PCA(n_components=2)
+    coords = pca.fit_transform(topic_term_matrix)
+
+    fig = px.scatter(
+        x=coords[:, 0],
+        y=coords[:, 1],
+        text=[f"Topic {i}" for i in range(len(coords))],
+        hover_name=[f"Topic {i}" for i in range(len(coords))],
+        labels={'x': 'PCA 1', 'y': 'PCA 2'}
+    )
+    fig.update_traces(marker=dict(size=12, color='indianred'), textposition='top center')
+    fig.update_layout(margin=dict(l=20, r=20, t=60, b=20))
+    return plot_to_div(fig)
+
+def create_interactive_bar_charts(lda_model):
+    data = []
+    for topic_id in range(lda_model.num_topics):
+        topic_words = lda_model.show_topic(topic_id, topn=10)
+        for word, weight in topic_words:
+            data.append({"Topic": f"Topic {topic_id}", "Word": word, "Weight": weight})
+
+    df = pd.DataFrame(data)
+    fig = px.bar(df, x='Weight', y='Word', color='Topic', orientation='h', height=600)
+    fig.update_layout(yaxis={'categoryorder': 'total ascending'})
+    return plot_to_div(fig)
+
+def create_interactive_heatmap(lda_model):
+    data = []
+    for topic_id in range(lda_model.num_topics):
+        for word, weight in lda_model.show_topic(topic_id, topn=10):
+            data.append({"Topic": f"Topic {topic_id}", "Word": word, "Weight": weight})
+
+    df = pd.DataFrame(data)
+    pivot_df = df.pivot(index="Word", columns="Topic", values="Weight").fillna(0)
+    fig = px.imshow(pivot_df, labels=dict(x="Topic", y="Word", color="Weight"))
+    return plot_to_div(fig)
+
+def create_interactive_topic_evolution(lda_model, corpus):
+    topic_matrix = np.zeros((len(corpus), lda_model.num_topics))
+    for i, doc in enumerate(corpus):
+        for topic_id, weight in lda_model.get_document_topics(doc):
+            topic_matrix[i, topic_id] = weight
+
+    df = pd.DataFrame(topic_matrix, columns=[f"Topic {i}" for i in range(lda_model.num_topics)])
+    df["Time"] = range(len(df))
+    fig = px.line(df, x="Time", y=df.columns[:-1])
+    return plot_to_div(fig)
+
+def create_interactive_topic_distribution(lda_model, corpus):
+    data = []
+    for doc_id, doc in enumerate(corpus):
+        for topic_id, weight in lda_model.get_document_topics(doc):
+            data.append({"Document": f"Doc {doc_id}", "Topic": f"Topic {topic_id}", "Weight": weight})
+
+    df = pd.DataFrame(data)
+    fig = px.bar(df, x="Document", y="Weight", color="Topic", height=600)
+    return plot_to_div(fig)
+
[…]

[…]
