### 2025-12-11 vinayakdasgupta/anvay#2 @vinayakdasgupta
Dear @x-tabdeveloping

I think I have managed to address the issues that you have raised.

1. I agree that topic interpretability is critical. In the updated version, all visualisations now display the top words for each topic on hover. This ensures that users do not need to switch back and forth between the Report and Visualization pages to understand what each topic represents. Also, all topics now have the same colour scheme set globally.
I would have prefered to have the hover on the legend, but that is, strangely, not possible for Plotly.

2. I have added a line in the documentation mentioning that both Heatmap and Bar Chart visualise the same topic–word weight matrix.

3. The tick issue was tricky. I have reduced the number of terms on display. This allows the user to see each term associated with the plot. I have also added the information on hover where it is on the x-axis (for instance in the topic distribution graph)

4. I have clarified the axis label. This is, indeed, the document upload order. I thought it might be useful if one were to look at, say, the corpus of an author arranged  by date (which, of course, has to reflect in the filename). I have also added the filename on hover, so the user can see which file they are looking at.

5. Merged clusters now display aggregated top keywords when hovered, making the hierarchical relationships much easier to interpret.

6. All visualisations use the same Roboto/Noto Bengali font stack as the rest of the site. Also, reduced Plotly margins.

7. I don't know how I forgot the spinner. It's now added.

Thanks so much for all of these suggestions. These make the visualizations far more intuitive.
