### 2025-12-11 b1cb175
Delete viz_old.py
