### 2025-12-11 1f657e3
updated paper
Updated paper.md with a State of the Field section.
