+++ paper.bib
@@ -1,12 +1,3 @@
-@article{blei2003latent,
-  title={Latent Dirichlet Allocation},
-  author={Blei, David M. and Ng, Andrew Y. and Jordan, Michael I.},
-  journal={Journal of Machine Learning Research},
-  volume={3},
-  pages={993--1022},
-  year={2003}
-}
-
 @inproceedings{rehurek_lrec,
       title = {{Software Framework for Topic Modelling with Large Corpora}},
       author = {Radim {\v R}eh{\r u}{\v r}ek and Petr Sojka},
@@ -20,14 +11,6 @@
       address = {Valletta, Malta},
       note={\url{http://is.muni.cz/publication/884893/en}},
       language={English}
-}
-
-@inproceedings{hagberg2008exploring,
-  title={Exploring Network Structure, Dynamics, and Function using NetworkX},
-  author={Hagberg, Aric and Swart, Pieter and Schult, Daniel A.},
-  booktitle={Proceedings of the 7th Python in Science Conference (SciPy)},
-  year={2008},
-  doi={10.25080/TCWV9851}
 }
 
 @inproceedings{chakrabarty-etal-2017-context,
@@ -56,13 +39,6 @@
   eprint={2107.03844}
 }
 
-@article{pyldavis2014,
-  title={pyLDAvis: Interactive topic model visualization},
-  author={Sievert, Carson and Shirley, Kenneth},
-  journal={ICML Workshop on Topic Models},
-  year={2014}
-}
-
 @misc{voyant,
   title={Voyant Tools},
   author={Sinclair, Stéfan and Rockwell, Geoffrey},
@@ -70,13 +46,193 @@
   howpublished={https://voyant-tools.org/}
 }
 
-@article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
-  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={Journal of Machine Learning Research},
-  volume={12},
-  pages={2825--2830},
-  year={2011},
-  publisher={JMLR. org}
+@inproceedings{bianchi-etal-2021-pre,
+    title = "Pre-training is a Hot Topic: Contextualized Document Embeddings Improve Topic Coherence",
+    author = "Bianchi, Federico  and
+      Terragni, Silvia  and
+      Hovy, Dirk",
+    editor = "Zong, Chengqing  and
+      Xia, Fei  and
+      Li, Wenjie  and
+      Navigli, Roberto",
+    booktitle = "Proceedings of the 59th Annual Meeting of the Association for Computational Linguistics and the 11th International Joint Conference on Natural Language Processing (Volume 2: Short Papers)",
+    month = aug,
+    year = "2021",
+    address = "Online",
+    publisher = "Association for Computational Linguistics",
+    url = "https://aclanthology.org/2021.acl-short.96/",
+    doi = "10.18653/v1/2021.acl-short.96",
+    pages = "759--766",
[…]
