+++ paper.md
@@ -20,39 +20,50 @@ bibliography: paper.bib
 
 # Summary
 
-*anvay* is a web-based tool for topic modelling in bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA) [@blei2003latent]—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, anvay presents the model output to be read, interpreted, and adjusted by the user.
+*anvay* is a web-based tool for topic modelling in Bengali, developed for exploratory reading and interpretive analysis. It provides a full pipeline for Latent Dirichlet Allocation (LDA) [@blei2003lda]—from corpus ingestion and preprocessing to model configuration and visual output—within a lightweight, browser-based interface. The tool foregrounds user interpretation: rather than providing coherence scores or fixed topic labels, anvay presents the model output to be read, interpreted, and adjusted by the user.
 
-Designed for literary, journalistic, and historical corpora in bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation, frequency filtering, and stopword pruning. The outputs, ranging from topic-word networks to document-level previews, are rendered with clarity and designed to enable close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
+Designed for literary, journalistic, and historical corpora in Bengali, *anvay* supports a range of language-specific preprocessing functions including lemmatisation, frequency filtering, and stopword pruning. The outputs, ranging from topic-word networks to document-level previews, are rendered with clarity and designed to enable close reading. Each topic is accessible through multiple lenses: top words, paragraph-level examples, document weights, and corpus-wide distribution.
 
 Beyond its technical function, *anvay* is an intervention in how we teach and understand computational methods within the humanities in low-resource contexts. 
 
+# State of the Field
+
+Topic modelling is well established in the digital humanities and is used for analysing large text collections in literary studies, cultural analytics, and media research [@jockers2013macroanalysis; @goldstone2014quiet]. Scholars employ models to identify recurring themes and support exploratory reading.
+
+Gensim [@rehurek_lrec] and Mallet [@mccallum2002mallet] are popular backend libraries. Several interfaces support this work, and *anvay* draws inspiration from many of these. Voyant Tools provides a general-purpose environment for text analysis, which includes an accessible topic-modelling component [@voyant]. Termite, one of the earliest contributions of this type, offers a clear tabular display for comparing topic–term relations [@chuang2012termite]. pyLDAvis supplies interactive views of topic distances and word relevance [@sievert-shirley-2014-ldavis]. TopicWizard presents topic clusters, keywords and document-level patterns through a web interface [@kardos2025topicwizardmodernmodelagnostic]. jsLDA demonstrates that browser-based topic modelling is feasible and useful in teaching contexts [@mimno_jslda]. Other platforms have made model outputs available to wider audiences. The Topic Modeling Tool provides a simple graphical front end to MALLET [@enderle2019topictool]. DFR-Browser supports exploratory reading of topic models within journal archives [@goldstone2014dfrbrowser]. 
+
+Recent systems such as BERTopic combine contextualised representations with neural topic models to improve coherence [@grootendorst2022bertopic]. Other methods, such as Top2Vec [@angelov2020top2vec] and CombinedTM [@bianchi-etal-2021-pre] built on sentence transformers or contextualised embeddings, also aim to improve topic coherence and reduce dependence on bag-of-words features.
+
+
 # Statement of Need
 
[…]

[…]
