### 2025-12-12 vinayakdasgupta/anvay#2 @x-tabdeveloping
I really like the updates to the app, it looks much more professional and it's more intuitive to use. I'm very satisfied!
