### 2025-12-17 138c6ee
Footer update
Updated footer to 1.1.1
