+++ results/topics.txt
@@ -1,30 +1,30 @@
 Topic 0:
-মহারাজবিক্রম, নম্বরে, খুড়োমহারাজ, কুমারের, অন্নদার, বিক্রমের, বিক্রম, নাশ্যামা, আত্মগত, প্রবেশবিক্রম
+কমলের, বদলাতে, বাগবাজারে, নিদেন, গয়লা, স্নায়ুর, আমা, মেয়েমানুষকে, পাঁচজনে, বাপমা
 
 Topic 1:
-ঠাকুরসন্ন্যাসী, কাননকুসুমিকা, লক্ষেশ্বর, নিবারণকে, দেখিচন্দ্রকান্ত, হবেচন্দ্রকান্ত, বেতসিনীর, প্রবেশলক্ষেশ্বর, প্রবেশচন্দ্রকান্ত, নিবারণবাবু
+শুন, রাজকার্যে, পিতৃব্য, কাঁদিতেছে, যেথা, কহো, দুর্যোধন, বাছাকে, হতেছে, পতি
 
 Topic 2:
-অবু, কেদারের, মনোরমা, অবিনাশের, ম্যানেজারের, ভায়া, মুকুন্দলালের, টম্, পিতামহের, কোর্তা
+ঠাকুরসন্ন্যাসী, ক্লাব, সন্ন্যাসীঠাকুর, ছোঁড়াটা, হ্যায়, ঋণশোধের, নেসন্ন্যাসী, বলেছিলেম, বেড়িয়েছি, মল্লিক
 
 Topic 3:
-বুড়া, পাঠাইতে, সীতারামকে, শেয়াল, বাড়িতেছে, বেসুরের, বহিতেছে, আচ্ছাদন, রাজবাটীতে, রমাইয়ের
+ব্রেস্, কনে, জন্মদিন, রোদ, খোকা, সর্দারকে, বিশু, পিণ্ড, লিখছ, সত্যযুগের
 
 Topic 4:
-শ্রীমতীর, বুদ্ধো, দেবদত্তের, রাও, বুদ্ধ, বুদ্ধের, কল্যাণী, নামালতী, বিনি, দুর্যোধন
+ভায়া, দেরাজ, রেখায়, বউরানীর, সাহিত্যিক, তালুক, দৈবের, পৌঁছোল, মহাজন, কাঁপছে
 
 Topic 5:
-অ্যাল্, ক্লাব, ব্রেস্, মিসেস, বিধু, সাহিত্যিক, কোনোটা, ড্, হরেন, সুকুমারী
+এনু, কহো, নরক, জননীর, স্নায়ুর, কমলের, জননী, দুর্যোধন, বদলাতে, শিরে
 
 Topic 6:
-যমুনার, ভাইফোঁটার, নাযতীন, ফরাসী, রুগীর, নেযতীন, বাস্তবকে, জীবনটাকে, স্ত্রীপুরুষের, ডাকাতের
+কীর্তন, শুভদিন, রাজাই, মিলব, মর্জি, নাচে, বোস্, দূতের, ঢেলা, পাগলা
 
 Topic 7:
-মুক্তধারার, অধৈর্য, পৌঁছোল, টেলিগ্রাফ, বউরানীর, তালুক, মুকুন্দলাল, দেরাজ, তক্তাপোশের, গোপাল
+ব্যাগ, প্রবেশবিপিন, প্রতিজ্ঞার, শৈলের, সভ্যের, প্রস্তাবটা, মধ্যম, ভগ্নীপতির, চন্দ্রবাবুকে, দড়িটা
 
 Topic 8:
[…]
