### 2025-12-17 fc400bb
Viz debug
Fixed issues:

- scatterplot and barchart display topic 0 only
- colours repeating -- now separated by colour space
