+++ viz.py
@@ -30,20 +30,40 @@ FONT_SIZE = 11
 BRAND_COLOR = "#0D85D8"
 DEFAULT_PALETTE = px.colors.qualitative.Plotly  # reliable qualitative palette
 
+import colorsys
+
 def build_topic_color_map(num_topics):
     """
-    Return a dict mapping 'Topic 0'..'Topic N' -> hex colour.
-    Ensures Topic 0 uses BRAND_COLOR and cycles the default palette for others.
+    Generate a deterministic, non-repeating colour for each topic.
+    Topic 0 uses the brand colour.
+    Remaining topics are evenly spaced in HSV colour space.
     """
-    # create working palette copy
-    palette = DEFAULT_PALETTE.copy()
-    # ensure first colour is brand colour
-    if len(palette) > 0:
-        palette[0] = BRAND_COLOR
-    # extend palette if num_topics exceeds length by repeating
-    color_list = [palette[i % len(palette)] for i in range(num_topics)]
-    return {f"Topic {i}": color_list[i] for i in range(num_topics)}
-# -----------------------------------------------------------------------
+
+    def hsv_to_hex(h, s, v):
+        r, g, b = colorsys.hsv_to_rgb(h, s, v)
+        return f"#{int(r*255):02x}{int(g*255):02x}{int(b*255):02x}"
+
+    color_map = {}
+
+    # Topic 0: brand colour
+    color_map["Topic 0"] = BRAND_COLOR
+
+    if num_topics <= 1:
+        return color_map
+
+    # Remaining topics: evenly spaced hues
+    for i in range(1, num_topics):
+        # distribute hues uniformly around the circle
+        hue = (i - 1) / max(1, (num_topics - 1))
+        # tuned for readability on white background
+        color_map[f"Topic {i}"] = hsv_to_hex(
+            h=hue,
+            s=0.65,
+            v=0.85
+        )
+
+    return color_map
+
 
 
 def get_topic_tooltips(lda_model, topn=5):
@@ -78,39 +98,32 @@ def create_interactive_scatter(lda_model):
     coords = pca.fit_transform(topic_term_matrix)
 
     tooltips = get_topic_tooltips(lda_model, topn=5)
-    labels = [f"Topic {i}" for i in range(len(coords))]
-
-    df = pd.DataFrame({
-        "x": coords[:, 0],
-        "y": coords[:, 1],
-        "label": labels,
-        "top_words": [tooltips[i] for i in range(len(coords))]
-    })
-
     color_map = build_topic_color_map(len(coords))
 
-    # Tell px to colour by the label column and use our map
-    fig = px.scatter(
-        df,
-        x="x",
-        y="y",
-        text="label",
-        color="label",                      # <- important
-        color_discrete_map=color_map,
-        height=480
-    )
-    fig.update_layout(font=dict(family=FONT_FAMILY, size=FONT_SIZE))
+    fig = go.Figure()
 
-    # only set marker size here — do NOT set marker.color
-    fig.update_traces(
-        marker=dict(size=14),
-        textposition='top center',
-        hovertemplate="<b>%{text}</b><br>Top words: %{customdata[0]}<extra></extra>",
-        customdata=df[["top_words"]].values
+    for topic_id, (x, y) in enumerate(coords):
+        label = f"Topic {topic_id}"
+        fig.add_trace(go.Scatter(
+            x=[x],
+            y=[y],
+            mode="markers+text",
+            text=[label],
+            textposition="top center",
+            marker=dict(size=14, color=color_map[label]),
+            customdata=[[tooltips.get(topic_id, "")]],
+            hovertemplate="<b>%{text}</b><br>Top words: %{customdata[0]}<extra></extra>"
+        ))
+
+    fig.update_layout(
+        height=480,
[…]
