+++ results/topics.txt
@@ -1,30 +1,90 @@
 Topic 0:
-কমলের, বদলাতে, বাগবাজারে, নিদেন, গয়লা, স্নায়ুর, আমা, মেয়েমানুষকে, পাঁচজনে, বাপমা
+কথা, হয়ে, করিয়া, দিয়ে, একটা, এক, হইয়া, যায়, নে, নি
 
 Topic 1:
-শুন, রাজকার্যে, পিতৃব্য, কাঁদিতেছে, যেথা, কহো, দুর্যোধন, বাছাকে, হতেছে, পতি
+বললে, দিয়ে, অমিত, একটা, হয়ে, কথা, এক, লাবণ্য, যায়, বললুম
 
 Topic 2:
-ঠাকুরসন্ন্যাসী, ক্লাব, সন্ন্যাসীঠাকুর, ছোঁড়াটা, হ্যায়, ঋণশোধের, নেসন্ন্যাসী, বলেছিলেম, বেড়িয়েছি, মল্লিক
+করিয়া, কথা, হয়ে, হইয়া, একটা, কহিলেন, এক, কহিল, ভালো, বলিয়া
 
 Topic 3:
-ব্রেস্, কনে, জন্মদিন, রোদ, খোকা, সর্দারকে, বিশু, পিণ্ড, লিখছ, সত্যযুগের
+কথা, হয়ে, একটা, ভাই, জন্যে, ভালো, একটু, রসিকবাবু, আপনাদের, নি
 
 Topic 4:
-ভায়া, দেরাজ, রেখায়, বউরানীর, সাহিত্যিক, তালুক, দৈবের, পৌঁছোল, মহাজন, কাঁপছে
+কথা, করিয়া, হয়ে, এক, কহিলেন, হইয়া, একটা, দিয়ে, বড়ো, নি
 
 Topic 5:
-এনু, কহো, নরক, জননীর, স্নায়ুর, কমলের, জননী, দুর্যোধন, বদলাতে, শিরে
+কথা, করিয়া, হয়ে, একটা, কহিলেন, বললে, দিয়ে, এক, তাঁহার, হইয়া
 
 Topic 6:
-কীর্তন, শুভদিন, রাজাই, মিলব, মর্জি, নাচে, বোস্, দূতের, ঢেলা, পাগলা
+হয়ে, কথা, একটা, দিয়ে, বললে, বললুম, নি, এক, বড়ো, নে
 
 Topic 7:
-ব্যাগ, প্রবেশবিপিন, প্রতিজ্ঞার, শৈলের, সভ্যের, প্রস্তাবটা, মধ্যম, ভগ্নীপতির, চন্দ্রবাবুকে, দড়িটা
+করিয়া, কথা, হইয়া, কহিলেন, একটা, হয়ে, দিয়ে, তাঁহার, কেবল, হইল
 
 Topic 8:
-শিখরে, গিনি, কমলালেবুর, তুলব, দুপুর, বসলেন, অর্কিড, প্রিয়া, হোটেল, আছেআমি
+কথা, করিয়া, হয়ে, দিয়ে, একটা, নি, বড়ো, একটু, বললে, নে
 
 Topic 9:
-প্রাতঃকাল, ধরিতে, পড়িলে, যাইতেছেন, শুনাইয়া, আচ্ছাদন, কেদারের, ফেলিলেন, ইহারা, পাইয়াছে
+কাজকরা, প্যাম্ফ্, হোটেলেই, আছেঅমিত, তাজমহলের, কথা, বললে, হয়ে, এক, দিয়ে
+
+Topic 10:
+একটা, কথা, করিয়া, দিয়ে, বড়ো, হয়ে, কহিলেন, নে, এক, বললে
+
+Topic 11:
[…]

[…]
