### 2025-12-17 360dd42
marker auto colour
Changed marker auto colour logic to golden-ratio hue jumping + lightness alternation
