+++ results/doc_topic_weights.txt
@@ -1,13 +1,13 @@
-ঘরে-বাইরে.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0, Topic 5: 0.0, Topic 6: 0.9955000281333923, Topic 7: 0.0, Topic 8: 0.0, Topic 9: 0.0, Topic 10: 0.0, Topic 11: 0.0, Topic 12: 0.0, Topic 13: 0.0, Topic 14: 0.0, Topic 15: 0.0, Topic 16: 0.0, Topic 17: 0.0, Topic 18: 0.0, Topic 19: 0.0, Topic 20: 0.0, Topic 21: 0.0, Topic 22: 0.0, Topic 23: 0.0, Topic 24: 0.0, Topic 25: 0.0, Topic 26: 0.0, Topic 27: 0.0, Topic 28: 0.0, Topic 29: 0.0
-বউ-ঠাকুরানীর হাট.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0, Topic 5: 0.0, Topic 6: 0.0, Topic 7: 0.0, Topic 8: 0.0, Topic 9: 0.0, Topic 10: 0.0, Topic 11: 0.0, Topic 12: 0.0, Topic 13: 0.0, Topic 14: 0.10750000178813934, Topic 15: 0.0, Topic 16: 0.0, Topic 17: 0.0, Topic 18: 0.0, Topic 19: 0.0, Topic 20: 0.0, Topic 21: 0.0, Topic 22: 0.0, Topic 23: 0.8924000263214111, Topic 24: 0.0, Topic 25: 0.0, Topic 26: 0.0, Topic 27: 0.0, Topic 28: 0.0, Topic 29: 0.0
-রাজর্ষি.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.0, Topic 4: 0.0, Topic 5: 0.0, Topic 6: 0.0, Topic 7: 0.0, Topic 8: 0.0, Topic 9: 0.0, Topic 10: 0.0, Topic 11: 0.0, Topic 12: 0.0, Topic 13: 0.0, Topic 14: 1.0, Topic 15: 0.0, Topic 16: 0.0, Topic 17: 0.0, Topic 18: 0.0, Topic 19: 0.0, Topic 20: 0.0, Topic 21: 0.0, Topic 22: 0.0, Topic 23: 0.0, Topic 24: 0.0, Topic 25: 0.0, Topic 26: 0.0, Topic 27: 0.0, Topic 28: 0.0, Topic 29: 0.0
-প্রজাপতির নির্বন্ধ.txt: Topic 0: 0.0, Topic 1: 0.0, Topic 2: 0.0, Topic 3: 0.9007999897003174, Topic 4: 0.0, Topic 5: 0.0, Topic 6: 0.0, Topic 7: 0.0, Topic 8: 0.0, Topic 9: 0.0, Topic 10: 0.0, Topic 11: 0.0, Topic 12: 0.0, Topic 13: 0.0, Topic 14: 0.0, Topic 15: 0.0, Topic 16: 0.0, Topic 17: 0.0, Topic 18: 0.0, Topic 19: 0.0, Topic 20: 0.0, Topic 21: 0.0, Topic 22: 0.0, Topic 23: 0.0, Topic 24: 0.0, Topic 25: 0.0, Topic 26: 0.0, Topic 27: 0.0, Topic 28: 0.0, Topic 29: 0.0
[…]

[…]
