### 2026-01-06 a012fb9
Updates and bug fixes
Topic weight consistency

Global topic prevalence is now computed once and reused consistently across visualisations.

Topic weight has been added to the scatterplot tooltip without altering the distance-based visual encoding.

Scatterplot interpretability

Topic weight is displayed on hover for each topic point.

The scatterplot remains distance-only in its visual encoding; weight is provided as inspectable metadata.

Axis semantics for the scatterplot are now explicitly documented, clarifying that both axes represent a two-dimensional projection used to visualise inter-
[…]
