+++ results/topic_words.txt
@@ -1,30 +1,6 @@
-Topic 0: কথা, করিয়া, একটা, হইয়া, হয়ে, কহিলেন, দিয়ে, বড়ো, তাঁহার, বললে
-Topic 1: কথা, করিয়া, একটা, হয়ে, হইয়া, দিয়ে, কহিলেন, এক, ভালো, তাঁহার
-Topic 2: হয়ে, কথা, একটা, করিয়া, দিয়ে, কেবল, জন্যে, নি, ভাই, হইয়া
-Topic 3: কথা, মহারাজ, রাজার, মহারানী, করো, কাশ্মীরের, তোমাদের, তোমাকে, রাজা, ভয়
-Topic 4: করিয়া, কথা, হয়ে, একটা, হইয়া, কহিলেন, এক, দিয়ে, কেবল, যায়
-Topic 5: কথা, হয়ে, দিয়ে, করিয়া, একটা, বড়ো, এক, কহিলেন, যায়, নি
-Topic 6: করিয়া, কহিলেন, হইয়া, তাঁহার, কথা, একটা, হয়ে, লইয়া, হইল, রাজা
-Topic 7: করিয়া, হয়ে, কথা, দিয়ে, একটা, বললে, নি, হইয়া, ভালো, সময়
-Topic 8: করিয়া, কহিলেন, কহিল, হইয়া, বিভা, কথা, রায়, তাঁহার, একটা, বলিয়া
-Topic 9: কথা, করিয়া, হয়ে, একটা, দিয়ে, কহিলেন, হইয়া, বললে, বড়ো, সময়
-Topic 10: কথা, হয়ে, একটা, ভালো, জন্যে, ভাই, করিয়া, নি, দিয়ে, একটু
-Topic 11: কথা, বিয়ে, ভাই, হয়ে, তোর, নে, একটা, তুই, ভালো, নি
-Topic 12: কথা, একটা, করিয়া, হয়ে, দিয়ে, হইয়া, কহিলেন, তোমাকে, এক, নি
-Topic 13: করিয়া, কথা, একটা, হইয়া, কহিলেন, তাঁহার, যায়, হয়ে, কহিল, বলিয়া
-Topic 14: করিয়া, একটা, কথা, হইয়া, বলিয়া, কহিলেন, কহিল, হইল, বলিল, দিয়া
-Topic 15: করিয়া, কহিলেন, হইয়া, তাঁহার, কহিল, কথা, বলিয়া, হইল, করিলেন, লাগিল
-Topic 16: বললে, দিয়ে, হয়ে, কথা, একটা, বললুম, নি, তোমাকে, নে, পড়ে
-Topic 17: হয়ে, কথা, দিয়ে, একটা, বললে, নি, বললুম, নে, জন্যে, এক
-Topic 18: একটা, হয়ে, করিয়া, কথা, বললে, দিয়ে, হইয়া, কহিলেন, নি, বড়ো
-Topic 19: কথা, করিয়া, একটা, হইয়া, কহিলেন, হয়ে, দিয়ে, তাঁহার, বললে, যায়
-Topic 20: কথা, হয়ে, করিয়া, হইয়া, একটা, দিয়ে, নে, বললে, কহিলেন, এক
-Topic 21: কথা, হয়ে, একটা, করিয়া, এক, নে, দিয়ে, যায়, হইয়া, কহিলেন
-Topic 22: করিয়া, হইয়া, একটা, বলিল, শচীশ, কথা, দামিনী, দিয়া, বলিয়া, হইল
[…]

[…]
