### 2026-01-08 #8641 @abhishektiwari
@vinayakdasgupta Re: Python version can we make sure `README.md` reports supported Python version with compatible Numpy correctly. Currently it says

> Python 3.9 or higher

In addition, I strongly suggest adding a Dockerfile here to avoid these issues.
