### 2026-01-08 8ba19b4
added Dockerfile
added Dockerfile for ease of installation
