+++ app.py
@@ -569,4 +569,5 @@ def handle_503(e):
 
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run(host='0.0.0.0', port=5000, debug=False)
+
