### 2026-01-08 15b1eef
added .gitignore
