+++ uploads/আচ্ছন্ন.txt
@@ -1,88 +1,0 @@
-লতার লাবণ্য যেন
-কচি কিশলয়ে ঘেরা,
-সুকুমার প্রাণ তার মাধুরীতে ঢেকেছে-
-কোমল মুকুলগুলি
-চারি দিকে আকুলিত
-তারি মাঝে প্রাণ যেন লুকিয়ে রেখেছে।
-ওরে যেন ভালো করে দেখা যায় না,
-আঁখি যেন ডুবে গিয়ে কূল পায় না।
-সাঁঝের আভা নেমে এল, জ্যোৎস্না পাশে ঘুমিয়ে প ' ল,
-ফুলের গন্ধ দেখতে এসেছে,
-তারাগুলি ঘিরে বসেছে।
-পূরবী রাগিণীগুলি
-দূর হতে চলে আসে
-ছুঁতে তারে হয় নাকো ভরসা-
-কাছে কাছে ফিরে ফিরে
-মুখপানে চায় তারা,
-যেন তারা মধুময়ী দুরাশা।
-ঘুমন্ত প্রাণেরে ঘিরে
-স্বপ্নগুলি ঘুরে ফিরে
-গাঁথে যেন আলোকের কুয়াশা,
-ঢেকে তারে আছে কত,
-চারি দিকে
-শত শত
-অনিমিষ নয়নের পিয়াসা।
-ওদের আড়াল থেকে
-আবছায়া দেখা যায়
-অতুলন প্রাণের বিকাশ,
-সোনার মেঘের মাঝে
-কচি উষা
-ফোটে ফোটে
-পুরবেতে তাহারি আভাস।
-আলোকবসনা যেন
-আপনি সে
-ঢাকা আছে
-আপনার রূপের মাঝার,
-রেখা রেখা হাসিগুলি
-আশেপাশে চমকিয়ে
-রূপেতেই লুকায় আবার।
-আঁখির আলোকছায়া
-আঁখিরে রয়েছে ঘিরে,
-তারি মাঝে দৃষ্টি পথহারা,
-যেথা চলে স্বর্গ হতে
-অবিরাম পড়ে যেন
-লাবণ্যের পুষ্পবারিধারা।
-ছবি ও গান
-ধরণীরে ছুঁয়ে যেন
-পা দুখানি ভেসে যায়,
-কুসুমের স্রোত বহে যায়,
-কুসুমেরে ফেলে রেখে
-খেলাধুলা ভুলে গিয়ে
-মায়ামুগ্ধ বসন্তের বায়।
-ওরে কিছু শুধাইলে
-বুঝি রে নয়ন মেলি
-দু দণ্ড নীরবে চেয়ে রবে,
-অতুল অধর দুটি
-ঈষৎ টুটিয়ে বুঝি
-অতি ধীরে দুটি কথা কবে।
-আমি কি বুঝি সে ভাষা,
-শুনিতে কি পাব বাণী
-সে যেন কিসের
-প্রতিধ্বনি —
-মধুর মোহের মতো
-যেমনি ছুঁইবে প্রাণ
-ঘুমায়ে সে পড়িবে অমনি।
-হৃদয়ের দূর হতে
-সে যেন রে কথা কয়
-তাই তার অতি মৃদুস্বর,
-বায়ুর হিল্লোলে তাই
-আকুল কুমুদসম
-কথাগুলি কাঁপে থর থর।
-কে তুমি গো উষাময়ী,
-আপন কিরণ দিয়ে
-আপনারে করেছ গোপন,
-রূপের সাগর-মাঝে
-কোথা তুমি ডুবে আছ
-একাকিনী লক্ষ্মীর মতন!
[…]
