+++ uploads/মুক্তধারা.txt
@@ -1,1164 +1,0 @@
-উত্তরকূট পার্বত্য প্রদেশ। সেখানকার উত্তরভৈরব-মন্দিরে যাইবার পথ। দূরে আকাশে একটা অভ্রভেদী লৌহযন্ত্রের মাথাটা দেখা যাইতেছে এবং তাহার অপরদিকে ভৈরবমন্দিরচূড়ার ত্রিশূল। পথের পার্শ্বে আমবাগানে রাজা রণজিতের শিবির। আজ অমাবস্যায় ভৈরবের মন্দিরে আরতি, সেখানে রাজা পদব্রজে যাইবেন, পথে শিবিরে বিশ্রাম করিতেছেন। তাঁহার সভার যন্ত্ররাজ বিভূতি বহুবৎসরের চেষ্টায় লৌহযন্ত্রের বাঁধ তুলিয়া মুক্তধারা ঝরনাকে বাঁধিয়াছেন। এই অসামান্য কীর্তিকে পুরস্কৃত করিবার উপলক্ষ্যে উত্তরকূটের সমস্ত লোক ভৈরব-মন্দির-প্রাঙ্গণে উৎসব করিতে চলিয়াছে। ভৈরব-মন্ত্রে দীক্ষিত সন্ন্যাসীদল সমস্ত দিন স্তবগান করিয়া বেড়াইতেছে। তাহাদের কাহারও হাতে ধূপাধারে ধূপ জ্বলিতেছে, কাহারও হাতে শঙ্খ, কাহারও ঘন্টা। গানের মাঝে মাঝে তালে তালে ঘন্টা বাজিতেছে।
-গান
-জয় ভৈরব, জয় শংকর,
-জয় জয় জয় প্রলয়ংকর,
-শংকর শংকর।
-জয় সংশয়ভেদন,
-জয় বন্ধনছেদন,
-জয় সংকটসংহর
-শংকর শংকর।
-[সন্ন্যাসীদল গাহিতে গাহিতে প্রস্থান করিল
-পূজার নৈবেদ্য লইয়া একজন বিদেশী পথিকের প্রবেশ
-উত্তরকূটের নাগরিককে সে প্রশ্ন করিল
-পথিক। আকাশে ওটা কী গড়ে তুলেছে? দেখতে ভয় লাগে।
-নাগরিক। জান না? বিদেশী বুঝি? ওটা যন্ত্র।
-পথিক। কিসের যন্ত্র?
-নাগরিক। আমাদের যন্ত্ররাজ বিভূতি পঁচিশ বছর ধরে যেটা তৈরি করছিল, সেটা ওই তো শেষ হয়েছে, তাই আজ উৎসব।
-পথিক। যন্ত্রের কাজটা কী?
-নাগরিক। মুক্তধারা ঝরনাকে বেঁধেছে।
-পথিক। বাবা রে। ওটাকে অসুরের মাথার মতো দেখাচ্ছে, মাংস নেই, চোয়াল ঝোলা। তোমাদের উত্তরকূটের শিয়রের কাছে অমন হাঁ করে দাঁড়িয়ে; দিনরাত্তির দেখতে দেখতে তোমাদের প্রাণপুরুষ যে শুকিয়ে কাঠ হয়ে যাবে।
-নাগরিক। আমাদের প্রাণপুরুষ মজবুত আছে, ভাবনা করো না।
-পথিক। তা হতে পারে, কিন্তু ওটা অমনতরো সূর্যতারার সামনে মেলে রাখবার জিনিস নয়, ঢাকা দিতে পারলেই ভালো হত।
[…]
