+++ uploads/মাতা.txt
@@ -1,62 +1,0 @@
-কুয়াশার জাল
-আবরি রেখেছে প্রাতঃকাল—
-সেইমতো ছিনু আমি কতদিন
-আত্মপরিচয়হীন
-।
-অস্পষ্ট স্বপ্নের মতো করেছিনু অনুভব
-কুমারীচাঞ্চল্যতলে আছিল যে সঞ্চিত গৌরব,
-যে
-নিরুদ্ধ আলোকের মুক্তির আভাস,
-অনাগত
-দেবতার আসন্ন আশ্বাস,
-পুষ্পকোরকের বক্ষে অগোচর ফলের মতন।
-তুই কোলে এলি
-যবে অমূল্য রতন,
-অপূর্ব
-প্রভাতরবি,
-আশার অতীত
-যেন প্রত্যাশার ছবি—
-লভিলাম
-আপনার পূর্ণতারে
-কাঙাল
-সংসারে।
-প্রাণের
-রহস্য সুগভীর
-অন্তরগুহায়
-ছিল স্থির,
-সে আজ বাহির হল দেহ লয়ে উন্মুক্ত আলোতে
-অন্ধকার
-হতে ;
-সুদীর্ঘকালের
-পথে
-চলিল
-সুদূর ভবিষ্যতে।
-যে আনন্দ আজি মোর শিরায় শিরায় বহে
-গৃহের কোণের তাহা নহে।
-আমার
-হৃদয় আজি পান্থশালা,
-প্রাঙ্গণে
-হয়েছে দীপ জ্বালা।
-হেথা
-কারে ডেকে আনিলাম
-অনাদিকালের পান্থ কিছুকাল করিবে বিশ্রাম।
-এ
-বিশ্বের যাত্রী যারা চলে অসীমের পানে
-আকাশে
-আকাশে নৃত্যগানে—
-আমার শিশুর
-মুখে কলকোলাহলে
-বীথিকা
-সে
-- যাত্রীর গান আমি শুনিব এ বক্ষতলে।
-অতিশয়
-নিকটের, দূরের তবু এ—
-আপন অন্তরে এল, আপনার কহে তো কভু এ।
-বন্ধনে
-দিয়েছে ধরা শুধু ছিন্ন করিতে বন্ধন ;
-আনন্দের ছন্দ টুটে উচ্ছ্বসিছে এ মোর ক্রন্দন।
-জননীর
-এ
-বেদনা, বিশ্বধরণীর
-সে যে আপনার ধন—
-না পারে রাখিতে নিজে, নিখিলেরে করে নিবেদন।

[…]
