+++ stopwords/verb_suffixes.txt
@@ -1,48 +1,0 @@
-ছি  
-ছে  
-লাম  
-তাম  
-বেন  
-েন  
-তে  
-ল  
-িল  
-বা  
-িত  
-িসে  
-েই  
-লি  
-ছ  
-ল  
-তিলা  
-তিলাম  
-তিলি  
-তিল  
-তিলেন  
-ত  
-তেম  
-তেছ  
-তেলি  
-তেছিল  
-তেছিলা  
-তিস  
-তামি  
-তেমি  
-তিলেই  
-তিবা  
-তিবে  
-তেমেন  
-সি  
-লান  
-ছিল  
-ছিলা  
-ছিলাম  
-ইলে  
-ইল  
-ইলো  
-ইলেন  
-ব  
-বি  
-বে  
-তাছে  
-তেছেন 
