+++ stopwords/noun_roots.txt
@@ -1,287 +1,0 @@
-মানুষ  
-পুরুষ  
-মহিলা  
-মেয়ে  
-ছেলে  
-মেয়ে  
-বাচ্চা  
-বাচ্চি  
-শিশু  
-বাবা  
-মা  
-ভাই  
-বোন  
-পিতা  
-মাতা  
-শিক্ষক  
-শিক্ষিকা  
-গুরু  
-অধ্যাপক  
-দরজা  
-কর্মী  
-অভিনেতা  
-অভিনেত্রী  
-প্রধান  
-রাজা  
-রাণী  
-দেবী  
-বিশ্ববিদ্যালয়  
-জ্ঞানী  
-বিদেশী  
-কন্যা  
-তিনজন  
-ছাত্রী  
-আধিকারিক  
-কর্মচারী  
-সন্তান  
-গ্রাম  
-শহর  
-নগর  
-পাড়া  
-বাজার  
-রাস্তা  
-গলি  
-এলাকা  
-দেশ  
-রাজধানী  
-সমুদ্র  
-নদী  
-জল  
-পাহাড়  
-বন  
-নদী  
-দ্বীপ  
-মাঠ  
-মন্দির  
-মসজিদ  
-চার্চ  
-গির্জা  
-মাকাম  
-মহল্লা  
-মহাসড়ক  
-মেলার  
-হাসপাতাল  
-স্কুল  
-কলেজ  
-বিশ্ববিদ্যালয়  
-বই  
-কাগজ  
-কলম  
-পেন্সিল  
-চশমা  
-ঘড়ি  
-টেবিল  
-প্যান  
-ব্যাগ  
-সেট  
-ফোন  
-ল্যাপটপ  
-টেলিভিশন  
-রেডিও  
-ক্যামেরা  
-গাড়ি  
-বাইক  
-মোটর  
-চেয়ার  
-পর্দা  
-অফিস  
-ঘর  
-কাপড়  
-খাতা  
-থালা  
-বাটি  
-ছাতা  
-বেলুন  
-ব্যালেন্স  
-কাঁচ  
-চামচ  
-প্লেট  
-কাপ  
-গ্লাস  
-ফল  
-সবজি  
-ভুট্টা  
-মাছ  
-মাংস  
-ভাত  
-রুটি  
-পানি  
-চা  
-কফি  
-ডিম  
-বিস্কুট  
-সিঙ্গারা  
-গাছ  
-ফুল  
-পাতা  
-শাখা  
-ফল  
-মৌমাছি  
-পিপঁড়ে  
-পাখি  
-জন্তু  
-অলিন্দ  
-সূর্য  
-চাঁদ  
-তারা  
-আকাশ  
-বৃষ্টি  
-পানি  
-মাটি  
-পাহাড়  
-নদী  
-বন  
-দ্বীপ  
-খাল  
-বাগান  
-সমুদ্র  
-ঝর্ণা  
-লেক  
-ভূমি  
-মেঘ  
-বেগ  
-হাওয়া  
-মাটি  
-রোদ  
-বাতাস  
-দিন  
-রাত  
-ঘণ্টা  
-মিনিট  
-সপ্তাহ  
-মাস  
-বছর  
-কাল  
-শুক্রবার  
-সোমবার  
-মঙ্গলবার  
-বুধবার  
-বৃহস্পতিবার  
-শুক্রবার  
-শনিবার  
-পর্যায়  
-ক্যালেন্ডার  
-ঋতু  
-অর্থাৎ  
-এখন  
-আগে  
-পরে  
-যাত্রা  
-সন্ধ্যা  
-প্রভাত  
-দুপুর  
-শীত  
[…]
