+++ stopwords/verb_roots.txt
@@ -1,152 +1,0 @@
-খাওয়া  
-দেখা  
-যাওয়া  
-আসা  
-করা  
-বলা  
-হওয়া  
-চলা  
-ধরা  
-পড়া  
-পাওয়া  
-সোয়া  
-খেলা  
-জানা  
-গোয়া  
-শোনা  
-লাগা  
-টাকা  
-বইয়া  
-হাসা  
-রেখা  
-নাচা  
-ভুলা  
-বসা  
-ঘোরা  
-ধ্বংস করা  
-গান গাওয়া  
-ঘুমানো  
-তুলা  
-বেরানো  
-পড়ানো  
-ভালোবাসা  
-চিন্তা করা  
-মজা নেওয়া  
-নখা  
-চেষ্টা করা  
-প্রশ্ন করা  
-মারা  
-নেওয়া  
-প্রস্তুত করা  
-জীবন যাপন করা  
-করানো  
-যানানো  
-খাওয়ানো  
-দেখানো  
-শোনানো  
-বোলানো  
-নাচানো  
-ফেলার  
-থামানো  
-শিক্ষানো  
-ভিক্ষা দেওয়া  
-গুছানো  
-চলা  
-যাওয়া  
-ফেরা  
-আসা  
-ঘোরানো  
-দৌড়ানো  
-ঝাঁপানো  
-পালানো  
-লাফানো  
-পালায়  
-বেড়ানো  
-যাত্রা করা  
-দেখা  
-শোনা  
-মাথা ঘুরানো  
-স্পর্শ করা  
-শ্বাস নেওয়া  
-বুঝা  
-চোখে দেখা  
-মাথা দেখা  
-দূরত্ব ঘোরা  
-চোখে পড়া  
-শুনানো  
-লিখা  
-বের করা  
-বিক্রি করা  
-খাবার দেওয়া  
-ফেলা  
-কাটা  
-ব্রেক করা  
-রেখে দেওয়া  
-সংগ্রহ করা  
-ঘুমানো  
-হাসা  
-ধরাল  
-নাচা  
-মাথা ধরে রাখা  
-ঝপকানো  
-গান গাওয়া  
-পড়া  
-যাচ্ছি  
-ওয়ালা  
-আছে  
-হচ্ছে  
-হয়  
-হয়ে  
-ছিল  
-হবে  
-মতো  
-ভেবে  
-দিতে  
-হতে  
-আসবে  
-হওয়া  
-প্রস্তুত হওয়া  
-চাই  
-আছে  
-হবে  
-এটা  
-করতে  
-করতে হবে  
-যাওয়া হবে  
-প্রস্তুত করা  
-ধরা  
-হওয়া  
-অবস্থান  
-আছ  
-রহা  
-পাওয়া  
-ছাড়া  
-ফেলে  
-দ্বারা  
-বাকি  
-চিন্তা  
-ভাবা  
-পছন্দ করা  
-মন্তব্য  
-প্রশ্ন করা  
-বিশ্বাস  
-সম্ভাবনা দেখা  
-মনোযোগ  
-লাগা  
-ভুল করা  
-বোলানো  
-খেলানো  
-শিখানো  
-ফেলা  
-তুলে নেয়া  
-ভ্রমণ করা  
-সংগ্রহ করা  
-নির্দেশনা দেওয়া  
-ঘড়ি নেওয়া  
-রাখা  
-ফেলা  
-সংগ্রহ করা  
-অংশগ্রহণ করা  
-কাজ করা  
-অধিকার লাভ করা  
-অভ্যস্ত হওয়া
