+++ results/topic_words.txt
@@ -1,6 +1,0 @@
-Topic 0: ঠাকুরসন্ন্যাসী, ঋণশোধের, ক্লাব, আগলে, গিনি, সন্ন্যাসীঠাকুর, পুঁতে, শয়তান, প্রিয়া, ভট্টাচার্য
-Topic 1: গোঁ, কলা, পৌঁছোল, দেরাজ, লণ্ঠন, দেহটা, কলার, কনে, তালুক, শব্দটা
-Topic 2: কেদারের, কাঁদিতেছে, সর্দারকে, নরক, করিছে, সেথা, সর্ব, শুন, গেলি, মালতী
-Topic 3: মজুমদার, গ্লাস, প্রতিজ্ঞার, ব্যাগ, শৈলের, সম্বন্ধেও, প্রবেশবিপিন, বিবাহযোগ্য, হোটেল, অর্কিড
-Topic 4: দেবদত্ত, পীতাম্বর, বুঝিলাম, জ্যাঠামশায়ের, নাস্তিক, রাজধানীতে, বদলাতে, শিবির, লিপ্ত, ঝরিয়া
-Topic 5: ব্রেস্, খোকা, জাম, সম্পাদক, ম্যাজিস্ট্রেটের, চাহে, ছড়াইয়া, দুঃখই, ক্রমশ, টুকরা
