+++ stopwords/noun_suffixes.txt
@@ -1,40 +1,0 @@
-টা  
-টি  
-গুলো  
-গুলা  
-গুলি  
-র  
-রা  
-দের  
-ের  
-ে  
-ার  
-েট  
-েটা  
-েটি  
-য়  
-য়ায়  
-তে  
-েতে  
-ায়  
-েয়  
-গণ  
-মণ্ডলী  
-বৃন্দ  
-াস  
-াদ  
-ালয়  
-ীক  
-ান  
-ারা  
-িরা  
-েক  
-হীন  
-াজ  
-াজী  
-ানী  
-ারী  
-িকা  
-েয়া  
-েয়া  
-ায়া  

[…]
