### 2026-01-10 59c3d1f
Extract document preprocessing from app.py
