### 2026-01-10 8944088
Extract corpus statistics computation from app.py
