### 2026-01-10 fc04ecf
Extract LDA training logic from app.py
