### 2026-01-10 b04d5bc
Introduce language-based preprocessing dispatcher
