### 2026-01-10 0bb7b8b
Enable English preprocessing pipeline
