+++ preprocessing/pipeline.py
@@ -1,4 +1,5 @@
 from preprocessing.bn import preprocess_bn
+from preprocessing.en import preprocess_en
 
 
 def preprocess_documents(
@@ -21,5 +22,17 @@ def preprocess_documents(
             percent,
             ngram
         )
+
+    elif language == "en":
+        return preprocess_en(
+            file_paths,
+            remove_stopwords,
+            custom_stopwords,
+            use_stemming,
+            normalisation_order,
+            percent,
+            ngram
+        )
+
     else:
         raise ValueError(f"Unsupported language: {language}")
