+++ preprocessing/pipeline.py
@@ -1,37 +1,72 @@
 from preprocessing.bn import preprocess_bn
 from preprocessing.en import preprocess_en
+from preprocessing.hi import preprocess_hi
+from preprocessing.ta import preprocess_ta
+from preprocessing.normalisation import (
+    resolve_normaliser,
+    resolve_default_normaliser,
+)
 
 
 def preprocess_documents(
     file_paths,
     remove_stopwords,
     custom_stopwords,
-    use_stemming,
+    normalisation,          # user request: "lemma" | "stem" | "none" | None
     normalisation_order,
     percent,
     ngram,
     language="bn"
 ):
+    """
+    Preprocess documents with language-specific logic and
+    resolved lexical normalisation.
+    """
+
+    # --- resolve normaliser centrally ---
+    if normalisation in (None, "auto"):
+        normaliser = resolve_default_normaliser(language)
+    else:
+        normaliser = resolve_normaliser(language, normalisation)
+
+    # --- dispatch to language-specific preprocessors ---
     if language == "bn":
         return preprocess_bn(
-            file_paths,
-            remove_stopwords,
-            custom_stopwords,
-            use_stemming,
-            normalisation_order,
-            percent,
-            ngram
+            file_paths=file_paths,
+            remove_stopwords=remove_stopwords,
+            custom_stopwords=custom_stopwords,
+            normaliser=normaliser,                 # ← NEW
+            normalisation_order=normalisation_order,
+            percent=percent,
+            ngram=ngram
         )
 
     elif language == "en":
         return preprocess_en(
-            file_paths,
-            remove_stopwords,
-            custom_stopwords,
-            use_stemming,
-            normalisation_order,
-            percent,
-            ngram
+            file_paths=file_paths,
+            remove_stopwords=remove_stopwords,
+            custom_stopwords=custom_stopwords,
+            normaliser=normaliser,                 # ← NEW
+            percent=percent,
+            ngram=ngram
+        )
+    elif language == "hi":
+        return preprocess_hi(
+            file_paths=file_paths,
+            remove_stopwords=remove_stopwords,
+            custom_stopwords=custom_stopwords,
+            normaliser=normaliser,
+            percent=percent,
+            ngram=ngram
+        )
+    elif language == "ta":
+        return preprocess_ta(
+            file_paths=file_paths,
+            remove_stopwords=remove_stopwords,
+            custom_stopwords=custom_stopwords,
+            normaliser=normaliser,
+            percent=percent,
+            ngram=ngram
         )
 
     else:
