+++ preprocessing/bn.py
@@ -15,26 +15,67 @@ def preprocess_bn(
     file_paths,
     remove_stopwords,
     custom_stopwords,
-    use_stemming,
+    normaliser,               # Normaliser or None
     normalisation_order,
     percent,
-    ngram,
-    language="bn"
+    ngram
 ):
-    # Build stopwords set
-    stop_words = set(stopwords.words('bengali')) if remove_stopwords else set()
+    """
+    Bengali preprocessing.
+
+    Supported normalisers:
+    - None
+    - Normaliser(key="bn_stem", kind="stem")
+    """
+
+    # ------------------------------------------------------------------
+    # Validate normaliser
+    # ------------------------------------------------------------------
+
+    if normaliser is not None and normaliser.key != "bn_stem":
+        raise ValueError(
+            f"Unsupported normaliser for Bengali: {normaliser.key}"
+        )
+
+    use_stemming = normaliser is not None
+
+    # ------------------------------------------------------------------
+    # Build stopword set
+    # ------------------------------------------------------------------
+
+    stop_words = set()
+
+    if remove_stopwords:
+        stop_words.update(stopwords.words("bengali"))
+
     if custom_stopwords:
         stop_words.update(custom_stopwords)
 
-    logging.info(f"Normalisation order: {normalisation_order}")
+    logging.info(
+        f"Stopword removal: {'enabled (NLTK Bengali)' if remove_stopwords else 'disabled'}"
+    )
 
-    all_tokens, raw_texts, doc_names = [], [], []
+    if use_stemming:
+        logging.info(
+            f"Lexical normalisation: bn_stem "
+            f"({normalisation_order.replace('_', ' → ')})"
+        )
+    else:
+        logging.info("Lexical normalisation: None")
+
+    # ------------------------------------------------------------------
+    # Process documents
+    # ------------------------------------------------------------------
+
+    all_tokens = []
+    raw_texts = []
+    doc_names = []
 
     for path in file_paths:
-        with open(path, 'r', encoding='utf-8') as f:
+        with open(path, "r", encoding="utf-8") as f:
             raw = f.read()
-            clean_text = remove_punctuation(raw)
 
+        clean_text = remove_punctuation(raw)
         tokens = custom_bengali_tokenize(clean_text)
 
         if use_stemming:
@@ -47,28 +88,34 @@ def preprocess_bn(
                 tokens = stem_tokens([tokens])[0]
 
             else:
-                # safe fallback: stem_first behaviour
+                # Defensive fallback: stem → filter
                 tokens = stem_tokens([tokens])[0]
                 tokens = [t for t in tokens if len(t) > 1 and t not in stop_words]
 
         else:
-            # stemming disabled → always filter only
+            # No normalisation → filter only
             tokens = [t for t in tokens if len(t) > 1 and t not in stop_words]
 
         all_tokens.append(tokens)
         raw_texts.append(clean_text)
         doc_names.append(os.path.basename(path))
 
-    # Optionally remove most frequent words
+    # ------------------------------------------------------------------
+    # Corpus-level frequency pruning
+    # ------------------------------------------------------------------
+
     if percent > 0:
         all_tokens = remove_most_frequent_words(all_tokens, percent)
 
-    # Build bigrams/trigrams if requested
-    if ngram == 'bigram':
+    # ------------------------------------------------------------------
+    # N-gram modelling
+    # ------------------------------------------------------------------
+
+    if ngram == "bigram":
         model = gensim.models.Phrases(all_tokens, min_count=2, threshold=50)
         all_tokens = [model[t] for t in all_tokens]
 
-    elif ngram == 'trigram':
+    elif ngram == "trigram":
[…]
