### 2026-01-17 #8641 @abhishektiwari
@vinayakdasgupta

"bengali" in paper title should be capitalised to "Bengali". Please update Zenodo title also.

Line 111–113: "While the system is tuned for formal Bengali prose, there are limitations: - Informal or dialectal orthographies may lead to malformed tokens - OCR artefacts or non-Unicode glyphs may interfere with tokenisation"  should be "While the system is tuned for formal Bengali prose, there are limitations: informal or dialectal orthographies may lead to malformed tokens, and OCR artefacts or non-Unicode glyphs may interfere with tokenisation."

**Issues with References:**

- Please add 10.25080/tcwv9851 as DOI for reference `Exploring Network Structure, Dynamics, and Function using NetworkX` in `paper.bib` file.
- `A review of bangla natural language processing tasks and the utility of transformer models` should be `A review of Bangla natural language processing tasks and the utility of transformer models`
- `Natural language processing with python` should be `Natural language processing with Python`
- `Inc., P. T. (2015). Collaborative data science. Plotly Technologies Inc.` parses corporate name as author name. Probably `author = {Plotly Technologies Inc.}` is incorrect and can be removed.
- `Latent dirichlet allocation` should be `Latent Dirichlet allocation`
- `Seaborn` should be `seaborn`
- `Topicwizard` should be `topicwizard`

By default JOSS converts reference title to title case please use extra {}s is in the case of proper names e.g., {P}ython, {B}angla.
