### 2026-01-19 dca4732
Updated paper.md and paper.bib
