+++ paper.bib
@@ -31,7 +31,7 @@
 }
 
 @article{alam2021review,
-  title={A Review of Bangla Natural Language Processing Tasks and the Utility of Transformer Models},
+  title={A Review of {B}angla Natural Language Processing Tasks and the Utility of Transformer Models},
   author={Alam, Firoj and Hasan, Md. Ashraful and Alam, Tanvir and Khan, Akib and Tajrin, Jannatul and Khan, Nayeem and Chowdhury, Shammur Absar},
   journal={arXiv preprint arXiv:2107.03844},
   year={2021},
@@ -40,7 +40,7 @@
 }
 
 @misc{voyant,
-  title={Voyant Tools},
+  title={{V}oyant Tools},
   author={Sinclair, Stéfan and Rockwell, Geoffrey},
   year={2016},
   howpublished={https://voyant-tools.org/}
@@ -80,7 +80,7 @@
 
 @inproceedings{chuang2012termite,
 author = {Chuang, Jason and Manning, Christopher D. and Heer, Jeffrey},
-title = {Termite: visualization techniques for assessing textual topic models},
+title = {{T}ermite: visualization techniques for assessing textual topic models},
 year = {2012},
 isbn = {9781450312875},
 publisher = {Association for Computing Machinery},
@@ -98,7 +98,7 @@ series = {AVI '12}
 
 @article{blei2003lda,
 author = {Blei, David M. and Ng, Andrew Y. and Jordan, Michael I.},
-title = {Latent dirichlet allocation},
+title = {{L}atent {D}irichlet {A}llocation},
 year = {2003},
 issue_date = {3/1/2003},
 publisher = {JMLR.org},
@@ -131,7 +131,7 @@ numpages = {30}
 }
 
 @misc{kardos2025topicwizardmodernmodelagnostic,
-      title={topicwizard -- a Modern, Model-agnostic Framework for Topic Model Visualization and Interpretation}, 
+      title={{t}opicwizard -- a Modern, Model-agnostic Framework for Topic Model Visualization and Interpretation}, 
       author={Márton Kardos and Kenneth C. Enevoldsen and Kristoffer Laigaard Nielbo},
       year={2025},
       eprint={2505.13034},
@@ -141,7 +141,7 @@ numpages = {30}
 }
 
 @misc{mimno_jslda,
-  title={jsLDA: In-Browser Topic Modeling},
+  title={js{LDA}: In-Browser Topic Modeling},
   author={Mimno, David},
   howpublished={\url{https://github.com/mimno/jsLDA}},
   note={Accessed 2025}
@@ -156,7 +156,7 @@ numpages = {30}
 }
 
 @misc{goldstone2014dfrbrowser,
-  title={DFR-Browser},
+  title={{DFR}-Browser},
   author={Goldstone, Andrew and Underwood, Ted},
   year={2014},
   howpublished={\url{https://github.com/agoldst/dfr-browser}},
@@ -164,7 +164,7 @@ numpages = {30}
 }
 
 @misc{grootendorst2022bertopic,
-      title={BERTopic: Neural topic modeling with a class-based TF-IDF procedure}, 
+      title={{BERT}opic: Neural topic modeling with a class-based TF-IDF procedure}, 
       author={Maarten Grootendorst},
       year={2022},
       eprint={2203.05794},
@@ -175,7 +175,7 @@ numpages = {30}
 
 
 @misc{angelov2020top2vec,
-      title={Top2Vec: Distributed Representations of Topics}, 
+      title={{T}op2{V}ec: Distributed Representations of Topics}, 
       author={Dimo Angelov},
       year={2020},
       eprint={2008.09470},
@@ -185,21 +185,11 @@ numpages = {30}
 }
 
 @misc{mccallum2002mallet,
[…]
