### 2026-01-20 #8641 @abhishektiwari
@vinayakdasgupta Looks like Plotly reference was removed from paper.bib but still referenced in paper.md. Can you fix this?
