### 2026-01-22 #8641 @abhishektiwari
Hi @vinayakdasgupta, following up on my last comment

> [@vinayakdasgupta](https://github.com/vinayakdasgupta) Looks like Plotly reference was removed from paper.bib but still referenced in paper.md. Can you fix this?
