### 2026-01-22 #8641 @vinayakdasgupta
Hi @abhishektiwari I fixed the plotly hook in the paper.md. I must have missed this earlier. Apologies for that. Hoping that this is all that is needed. Thank you again for all your help. Really appreciate it!
