### 2026-01-22 #8641 @vinayakdasgupta
@abhishektiwari Also, wanted to leave a note. I have had several peer-reviews in my academic career, but this remains my favourite. The ability to speak with the reviewer to express one's viewpoint, to improve the work based on their suggestions, and to be able to do it in an open and transparent way is what all peer reviews should aspire to. Unfortunately, in the humanities, this is rarely the case. I wanted to thank you and JOSS for facilitating such a fantastic review process.
