### 2026-01-22 #8641 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
