### 2026-01-31 #8641 @samhforbes
@vinayakdasgupta
- line 10 anvay should be in italics
- line 118 new line after GitHub so the whole URL displays
- Check formatting of references - you have some which (eg) are title case for the article / presentation name and some which are sentence case.
