### 2026-02-17 #8641 @vinayakdasgupta
Dear @samhforbes thank you so much! I am delighted. I have one concern -- my name is not displaying correctly in the citation: My first name is Vinayak and family name is Das Gupta. It is taking "Gupta" as my family name, which is not correct. Could this be addressed? Many thanks!
