### 2026-02-17 #8641 @samhforbes
Hi there, good point. Can you use the given-names and surname fields instead of name?
