### 2026-02-19 #8641 @vinayakdasgupta
Dear @samhforbes  I have used given-names and surname in the updated paper.md file. Many thanks!
